% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcdg_chl_1.R
\docType{data}
\name{gcdg_chl_1}
\alias{gcdg_chl_1}
\title{Chili, 0.5-1.75 years: \code{gcdg_chl_1}}
\format{
A \code{data.frame} with 2139 rows and 113 variables:\tabular{lll}{
   Name \tab Type \tab Label \cr
   \code{ctrcd} \tab chr \tab Country code (ISO 3166-1 alpha-3) \cr
   \code{cohort} \tab chr \tab Cohort name \cr
   \code{cohortn} \tab chr \tab Cohort number \cr
   \code{subjid} \tab int \tab Child number \cr
   \code{agedays} \tab int \tab Age (days) \cr
   \code{sex} \tab chr \tab Either \code{"male"} or  \code{"female"} \cr
   \code{gagebrth} \tab int \tab Gestational age (days) \cr
   \code{by1mdd045} \tab 0/1 \tab inspects own hand \cr
   \code{by1mdd046} \tab 0/1 \tab Closes on dangling ring (check hand preference) \cr
   \code{by1mdd047} \tab 0/1 \tab turns head to sound of bell \cr
   \code{by1mdd048} \tab 0/1 \tab turns head to sound of rattle \cr
   \code{by1mdd049} \tab 0/1 \tab reaches for cube \cr
   \code{by1mdd050} \tab 0/1 \tab manipulates table edge actively \cr
   \code{by1mdd051} \tab 0/1 \tab eye-hand coordination in reaching \cr
   \code{by1mdd052} \tab 0/1 \tab regards pellet \cr
   \code{by1mdd053} \tab 0/1 \tab mirror image approach \cr
   \code{by1mdd054} \tab 0/1 \tab picks up cube (check hand preference) \cr
   \code{by1mdd055} \tab 0/1 \tab Vocalises attitudes \cr
   \code{by1mdd056} \tab 0/1 \tab retains 2 cubes \cr
   \code{by1mdd057} \tab 0/1 \tab exploitive paper play \cr
   \code{by1mdd058} \tab 0/1 \tab Discriminates strangers \cr
   \code{by1mdd059} \tab 0/1 \tab recovers rattle, in crib \cr
   \code{by1mdd060} \tab 0/1 \tab reaches persistently \cr
   \code{by1mdd061} \tab 0/1 \tab Likes frolic play \cr
   \code{by1mdd062} \tab 0/1 \tab turns head after fallen spoon \cr
   \code{by1mdd063} \tab 0/1 \tab lifts inverted cup \cr
   \code{by1mdd064} \tab 0/1 \tab reaches for second cube \cr
   \code{by1mdd065} \tab 0/1 \tab smiles at mirror (5.4 months) \cr
   \code{by1mdd066} \tab 0/1 \tab bangs in play \cr
   \code{by1mdd067} \tab 0/1 \tab sustained inspection of ring \cr
   \code{by1mdd068} \tab 0/1 \tab exploitive string play \cr
   \code{by1mdd069} \tab 0/1 \tab transfers objects hand to hand \cr
   \code{by1mdd070} \tab 0/1 \tab picks up cube deftly and directly \cr
   \code{by1mdd071} \tab 0/1 \tab pulls string: secures ring \cr
   \code{by1mdd072} \tab 0/1 \tab interest in sound production \cr
   \code{by1mdd073} \tab 0/1 \tab lifts cup with handle \cr
   \code{by1mdd074} \tab 0/1 \tab attends to scribbling \cr
   \code{by1mdd075} \tab 0/1 \tab looks for fallen spoon \cr
   \code{by1mdd076} \tab 0/1 \tab playful response to mirror \cr
   \code{by1mdd077} \tab 0/1 \tab retains 2 of 3 cubes offered \cr
   \code{by1mdd078} \tab 0/1 \tab manipulates bell: insterest in details \cr
   \code{by1mdd079} \tab 0/1 \tab vocalizes 4 different syllables \cr
   \code{by1mdd080} \tab 0/1 \tab pulls string adaptively: secures ring \cr
   \code{by1mdd081} \tab 0/1 \tab cooperates in games (is this some other scale?) \cr
   \code{by1mdd082} \tab 0/1 \tab attempts to secure 3 cubes \cr
   \code{by1mdd083} \tab 0/1 \tab rings bell purposely \cr
   \code{by1mdd084} \tab 0/1 \tab listens selectively with familiar words \cr
   \code{by1mdd085} \tab 0/1 \tab says "da-da" or equivalent \cr
   \code{by1mdd086} \tab 0/1 \tab Uncovers toy (check manual if by pulling cloth \cr
   \code{by1mdd087} \tab 0/1 \tab fingers holes in pegboard \cr
   \code{by1mdd088} \tab 0/1 \tab picks up cup, secures cube \cr
   \code{by1mdd089} \tab 0/1 \tab responds to verbal request \cr
   \code{by1mdd090} \tab 0/1 \tab puts cube in cup on command \cr
   \code{by1mdd091} \tab 0/1 \tab looks for content of box \cr
   \code{by1mdd092} \tab 0/1 \tab stirs with spoon in imitation \cr
   \code{by1mdd093} \tab 0/1 \tab looks at pictures in book \cr
   \code{by1mdd094} \tab 0/1 \tab inhibits on command \cr
   \code{by1mdd095} \tab 0/1 \tab attempting to imitate scribble \cr
   \code{by1mdd096} \tab 0/1 \tab unwraps cube \cr
   \code{by1mdd097} \tab 0/1 \tab repeats performance laughed at \cr
   \code{by1mdd098} \tab 0/1 \tab holds crayon adaptively \cr
   \code{by1mdd099} \tab 0/1 \tab pushes car along \cr
   \code{by1mdd100} \tab 0/1 \tab puts 3 or more cubes in cup \cr
   \code{by1mdd101} \tab 0/1 \tab jabbers expressively \cr
   \code{by1mdd102} \tab 0/1 \tab uncovers blue box \cr
   \code{by1mdd103} \tab 0/1 \tab turns pages of books \cr
   \code{by1mdd104} \tab 0/1 \tab Pats toy (whistle doll) in imitation \cr
   \code{by1mdd105} \tab 0/1 \tab dangles ring by string \cr
   \code{by1mdd106} \tab 0/1 \tab imitates words \cr
   \code{by1mdd107} \tab 0/1 \tab puts beads in box (6 of 8) \cr
   \code{by1mdd108} \tab 0/1 \tab places 1 peg repeateadly \cr
   \code{by1mdd109} \tab 0/1 \tab removes pellet from bottle \cr
   \code{by1mdd110} \tab 0/1 \tab blue board: places 1 round block \cr
   \code{by1mdd111} \tab 0/1 \tab builds tower of 2 cubes \cr
   \code{by1mdd112} \tab 0/1 \tab scribbles sponstaneously \cr
   \code{by1mdd113} \tab 0/1 \tab says 2 words \cr
   \code{by1mdd114} \tab 0/1 \tab puts 9 cubes in cup \cr
   \code{by1mdd115} \tab 0/1 \tab closes round box \cr
   \code{by1mdd116} \tab 0/1 \tab uses gestures to  make wants known \cr
   \code{by1mdd117} \tab 0/1 \tab shows shoes or other clothing, or own toy \cr
   \code{by1mdd118} \tab 0/1 \tab pegs placed in 70 seconds \cr
   \code{by1mdd119} \tab 0/1 \tab builds tower of 3 cubes \cr
   \code{by1mdd120} \tab 0/1 \tab pink board: places round block \cr
   \code{by1mdd121} \tab 0/1 \tab blue board: places 2 round blocks \cr
   \code{by1mdd122} \tab 0/1 \tab attains toy with stick \cr
   \code{by1mdd123} \tab 0/1 \tab pegs placed in 42 seconds \cr
   \code{by1mdd124} \tab 0/1 \tab names 1 object \cr
   \code{by1mdd125} \tab 0/1 \tab imitates crayon stroke \cr
   \code{by1mdd126} \tab 0/1 \tab follows directions doll (check parts passed) \cr
   \code{by1mdd127} \tab 0/1 \tab uses words to  make wants known \cr
   \code{by1mdd128} \tab 0/1 \tab points to parts of doll (check parts recognised, check manua \cr
   \code{by1mdd129} \tab 0/1 \tab blue board: places 2 round and 2 square blocks \cr
   \code{by1mdd130} \tab 0/1 \tab names 1 picture \cr
   \code{by1mdd131} \tab 0/1 \tab finds 2 objects \cr
   \code{by1mdd132} \tab 0/1 \tab points to 3 pictures \cr
   \code{by1mdd133} \tab 0/1 \tab No label:  b1m133 \cr
   \code{by1mdd134} \tab 0/1 \tab pegs placed in 30 seconds \cr
   \code{by1mdd135} \tab 0/1 \tab differentiates scribble from stroke \cr
   \code{by1mdd136} \tab 0/1 \tab sentence of 2 words \cr
   \code{by1mdd137} \tab 0/1 \tab pink board: completes \cr
   \code{by1mdd138} \tab 0/1 \tab names 2 objects \cr
   \code{by1mdd139} \tab 0/1 \tab points to 5 pictures \cr
   \code{by1mdd140} \tab 0/1 \tab broken doll: mends approximately \cr
   \code{by1mdd141} \tab 0/1 \tab names 3 pictures \cr
   \code{by1mdd142} \tab 0/1 \tab blue board: places 6 blocks \cr
   \code{by1mdd143} \tab 0/1 \tab builds tower of 6 cubes \cr
   \code{by1mdd144} \tab 0/1 \tab discriminates 2: cup, plate, box \cr
   \code{by1mdd145} \tab 0/1 \tab names watch, 4th picture \cr
   \code{by1mdd146} \tab 0/1 \tab names 3 objects \cr
   \code{by1mdd150} \tab 0/1 \tab names watch, 2nd picture \cr
   \code{by1mdd151} \tab 0/1 \tab pink board: reverse \cr
   \code{by1mdd152} \tab 0/1 \tab discriminates 3: cup, plate, box \cr
   \code{by1mdd153} \tab 0/1 \tab broken doll: mends exactly \cr
}
}
\usage{
gcdg_chl_1
}
\description{
A dataset with developmental scores at the item level for
2139 unique children measured in the years 1991-1996 at ages clustered
around six months (n = 128), 12 months (n = 1732) and 18 month (n = 279).
}
\details{
Instruments: Bayley I (\code{by1})
}
\references{
Lozoff B, De Andraca I, Castillo M, Smith JB, Walter T, Pino P.
Behavioral and developmental effects of preventing iron-deficiency anemia
in healthy full-term infants. Pediatrics. 2003;112:846-854.
\url{https://pediatrics.aappublications.org/content/112/4/846.short}
}
\keyword{datasets}
