% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcdg_col_lt42m.R
\docType{data}
\name{gcdg_col_lt42m}
\alias{gcdg_col_lt42m}
\title{Colombia, 0.5-3.5 years: \code{gcdg_col_lt42m}}
\format{
A \code{data.frame} with 1311 rows and 627 variables:\tabular{lll}{
   Name \tab Type \tab Label \cr
   \code{ctrcd} \tab chr \tab Country code (ISO 3166-1 alpha-3) \cr
   \code{cohort} \tab chr \tab Cohort name \cr
   \code{cohortn} \tab chr \tab Cohort number \cr
   \code{subjid} \tab int \tab Child number \cr
   \code{agedays} \tab int \tab Age (days) \cr
   \code{sex} \tab chr \tab Either \code{"male"} or  \code{"female"} \cr
   \code{gagebrth} \tab int \tab Gestational age (days) \cr
   \code{aqicmc010} \tab 0/1 \tab When a loud noise occurs, does your baby turn to see where t \cr
   \code{aqicmc013} \tab 0/1 \tab Does your baby respond to the tone of your voice and stop he \cr
   \code{aqicmc014} \tab 0/1 \tab Does your baby make sounds like "da," "ga," "ka," and "ba"? \cr
   \code{aqicmc015} \tab 0/1 \tab If you copy the sounds your baby makes, does your baby repea \cr
   \code{aqicmc016} \tab 0/1 \tab Does your baby make two similar sounds like "ba-ba," "da-da, \cr
   \code{aqicmc018} \tab 0/1 \tab If you ask your baby to, does he play at least one nursery g \cr
   \code{aqicmc020} \tab 0/1 \tab When you ask, "Where is the ball (hat, shoe, etc.)?" does yo \cr
   \code{aqicmc021} \tab 0/1 \tab Does your baby follow one simple command, such as "Come here \cr
   \code{aqicmc022} \tab 0/1 \tab Does your baby say three words such as "Mama", "Dada" and "B \cr
   \code{aqicmc023} \tab 0/1 \tab When your child wants something, does she tell you by pointi \cr
   \code{aqicmc024} \tab 0/1 \tab Does your baby shake his head when he means "no" or "yes"? \cr
   \code{aqicmc025} \tab 0/1 \tab Does your baby/child say four or more words in addition to " \cr
   \code{aqicmc026} \tab 0/1 \tab Without your giving him clues by pointing or using gestures, \cr
   \code{aqicmc027} \tab 0/1 \tab When you ask her to, does your baby go into another room to \cr
   \code{aqicmc028} \tab 0/1 \tab When you ask your child to point to her nose, eyes, hair, fe \cr
   \code{aqicmc029} \tab 0/1 \tab Without your showing him, does your child point to the corre \cr
   \code{aqicmc030} \tab 0/1 \tab Does your child say eight or more words in addition to "Mama \cr
   \code{aqicmc031} \tab 0/1 \tab If you point to a picture of a ball (kitty, cup, hat, etc.) \cr
   \code{aqicmc032} \tab 0/1 \tab Without giving your child help by pointing or using gestures \cr
   \code{aqicmc033} \tab 0/1 \tab Does your child say 15 or more words in addition to "Mama" a \cr
   \code{aqicmc034} \tab 0/1 \tab Show your child how a zipper on a coat moves up and down, an \cr
   \code{aqicmc035} \tab 0/1 \tab Does your child imitate a two-word sentence? For example, wh \cr
   \code{aqicmc036} \tab 0/1 \tab Does your child say two or three words that represent differ \cr
   \code{aqicmc037} \tab 0/1 \tab When looking at a picture book, does your child tell you wha \cr
   \code{aqicmc038} \tab 0/1 \tab When you ask, "What is your name?" does your child say her f \cr
   \code{aqicmc039} \tab 0/1 \tab Does your child correctly use at least two words like "me," \cr
   \code{aqicmc040} \tab 0/1 \tab Does your child make sentences that are three or four words \cr
   \code{aqicmc041} \tab 0/1 \tab Without your giving help by pointing or repeating directions \cr
   \code{aqicmc055} \tab 0/1 \tab Does your child use all of the words in a sentence (for exam \cr
   \code{aqicmc061} \tab 0/1 \tab When you ask, "What is your name?" does your child say both \cr
   \code{aqifmc011} \tab 0/1 \tab Does your baby pick up a small toy with only one hand? \cr
   \code{aqifmc013} \tab 0/1 \tab Does your baby pick up a small toy, holding it in the center \cr
   \code{aqifmc014} \tab 0/1 \tab Does your baby reach for a crumb or Cheerio and touch it wit \cr
   \code{aqifmc015} \tab 0/1 \tab Does your baby try to pick up a crumb or Cheerio by using hi \cr
   \code{aqifmc016} \tab 0/1 \tab Does your baby successfully pick up a crumb or Cheerio by us \cr
   \code{aqifmc017} \tab 0/1 \tab Does your baby pick up a small toy with the tips of her thum \cr
   \code{aqifmc018} \tab 0/1 \tab After one or two tries, does your baby pick up a piece of st \cr
   \code{aqifmc019} \tab 0/1 \tab Does your baby pick up a crumb or Cheerio with the tips of h \cr
   \code{aqifmc020} \tab 0/1 \tab Does your baby put a small toy down, without dropping it, an \cr
   \code{aqifmc021} \tab 0/1 \tab Without resting his arm or hand on the table, does your baby \cr
   \code{aqifmc022} \tab 0/1 \tab Does your baby/child help turn the pages of a book? (You may \cr
   \code{aqifmc023} \tab 0/1 \tab Does your child turn the pages of a book by himself? (He may \cr
   \code{aqifmc024} \tab 0/1 \tab Does your baby/child make a mark on the paper with the tip o \cr
   \code{aqifmc025} \tab 0/1 \tab Does your baby/child stack a small block or toy on top of an \cr
   \code{aqifmc027} \tab 0/1 \tab Does your baby/child throw a small ball with a forward arm m \cr
   \code{aqifmc029} \tab 0/1 \tab Does your child get a spoon into her mouth right side up so \cr
   \code{aqifmc030} \tab 0/1 \tab Does your child stack six small blocks or toys on top of eac \cr
   \code{aqifmc031} \tab 0/1 \tab Does your child use a turning motion with her hand while try \cr
   \code{aqifmc032} \tab 0/1 \tab Does your child turn pages in a book, one page at a time? \cr
   \code{aqifmc033} \tab 0/1 \tab After your child watches you draw a single circle, ask him t \cr
   \code{aqifmc034} \tab 0/1 \tab After your child watches you draw a line from the top of the \cr
   \code{aqifmc035} \tab 0/1 \tab After your child watches you draw a line from one side of th \cr
   \code{aqifmc036} \tab 0/1 \tab Can your child string small items such as beads, macaroni or \cr
   \code{aqifmc037} \tab 0/1 \tab Does your child try to cut paper with child-safe scissors? S \cr
   \code{aqifmc038} \tab 0/1 \tab When drawing, does your child hold a pencil, crayon, or pen \cr
   \code{aqifmc039} \tab 0/1 \tab Does your child put together a five to seven piece interlock \cr
   \code{aqifmc041} \tab 0/1 \tab Using the shape at right to look at, does your child copy it \cr
   \code{aqigmc012} \tab 0/1 \tab When you put your baby on the floor, does she lean on her ha \cr
   \code{aqigmc013} \tab 0/1 \tab If you hold both hands just to balance your baby, does he su \cr
   \code{aqigmc015} \tab 0/1 \tab Does your baby roll from his back to his tummy, getting both \cr
   \code{aqigmc016} \tab 0/1 \tab When sitting on the floor, does your baby sit up straight fo \cr
   \code{aqigmc017} \tab 0/1 \tab Does your baby get into a crawling position by getting up on \cr
   \code{aqigmc018} \tab 0/1 \tab When you stand your baby next to furniture or the crib rail, \cr
   \code{aqigmc020} \tab 0/1 \tab While holding onto furniture, does your baby lower himself w \cr
   \code{aqigmc021} \tab 0/1 \tab While holding onto furniture, does your baby bend down and p \cr
   \code{aqigmc022} \tab 0/1 \tab Does your baby walk beside furniture while holding on with o \cr
   \code{aqigmc023} \tab 0/1 \tab When you hold one hand just to balance your baby, does she t \cr
   \code{aqigmc024} \tab 0/1 \tab Does your baby/child bend over or squat to pick up an object \cr
   \code{aqigmc025} \tab 0/1 \tab Does your baby/child climb onto furniture or other large obj \cr
   \code{aqigmc026} \tab 0/1 \tab Does your baby/child stand up in the middle of the floor by \cr
   \code{aqigmc027} \tab 0/1 \tab Does your baby/child move around by walking, rather than by \cr
   \code{aqigmc028} \tab 0/1 \tab Does your child walk well and seldom fall? \cr
   \code{aqigmc029} \tab 0/1 \tab Does your child climb on an object such as a chair to reach \cr
   \code{aqigmc030} \tab 0/1 \tab When you show your child how to kick a large ball, does he t \cr
   \code{aqigmc031} \tab 0/1 \tab Does your child walk either up or down at least two steps by \cr
   \code{aqigmc032} \tab 0/1 \tab Does your child walk down stairs if you hold onto one of her \cr
   \code{aqigmc033} \tab 0/1 \tab Does your child run fairly well, stopping herself without bu \cr
   \code{aqigmc034} \tab 0/1 \tab Without holding onto anything for support, does your child k \cr
   \code{aqigmc036} \tab 0/1 \tab Does your child jump with both feet leaving the floor at the \cr
   \code{aqigmc037} \tab 0/1 \tab While standing, does your child throw a ball overhand by rai \cr
   \code{aqigmc038} \tab 0/1 \tab Does your child walk up stairs, using only one foot on each \cr
   \code{aqigmc039} \tab 0/1 \tab Does your child jump forward at least 3 inches with both fee \cr
   \code{aqigmc040} \tab 0/1 \tab Does your child stand on one foot for about 1 second without \cr
   \code{aqigmc043} \tab 0/1 \tab Does your child catch a large ball with both hands? (You sho \cr
   \code{aqigmc044} \tab 0/1 \tab Does your child jump forward at least 6 inches with both fee \cr
   \code{aqipxc011} \tab 0/1 \tab When your baby is on her back, does she try to get a toy she \cr
   \code{aqipxc012} \tab 0/1 \tab Does your baby play by banging a toy up and down on the floo \cr
   \code{aqipxc014} \tab 0/1 \tab Does your baby pass a toy back and forth from one hand to th \cr
   \code{aqipxc015} \tab 0/1 \tab When holding a toy in his hand, does your baby bang it again \cr
   \code{aqipxc016} \tab 0/1 \tab Does your baby pick up two small toys, one in each hand, and \cr
   \code{aqipxc017} \tab 0/1 \tab After watching you hide a small toy under a piece of paper o \cr
   \code{aqipxc018} \tab 0/1 \tab While holding a small toy in each hand, does your baby clap \cr
   \code{aqipxc019} \tab 0/1 \tab Does your baby poke at or try to get a crumb or Cheerio that \cr
   \code{aqipxc020} \tab 0/1 \tab If you put a small toy into a bowl or box, does your baby co \cr
   \code{aqipxc021} \tab 0/1 \tab Does your baby drop two small toys, one after the other, int \cr
   \code{aqipxc022} \tab 0/1 \tab If you give your child a bottle, spoon, or pencil upside dow \cr
   \code{aqipxc023} \tab 0/1 \tab After you scribble back and forth on paper with a crayon (or \cr
   \code{aqipxc024} \tab 0/1 \tab Does your baby/child drop several small toys, one after anot \cr
   \code{aqipxc025} \tab 0/1 \tab Can your baby/child drop a crumb or Cheerio into a small, cl \cr
   \code{aqipxc026} \tab 0/1 \tab Without your showing him how, does your child scribble back \cr
   \code{aqipxc027} \tab 0/1 \tab After a crumb or Cheerio is dropped into a small, clear bott \cr
   \code{aqipxc028} \tab 0/1 \tab After a crumb or Cheerio is dropped into a small, clear bott \cr
   \code{aqipxc030} \tab 0/1 \tab Does your child put things away where they belong? For examp \cr
   \code{aqipxc031} \tab 0/1 \tab Does your child pretend objects are something else? For exam \cr
   \code{aqipxc032} \tab 0/1 \tab While your child watches, line up four objects like blocks o \cr
   \code{aqipxc033} \tab 0/1 \tab If your child wants something she cannot reach, does she fin \cr
   \code{aqipxc034} \tab 0/1 \tab After you have shown your baby  how, does he try to get a sm \cr
   \code{aqipxc035} \tab 0/1 \tab When looking in the mirror, ask "Where is _______?" (Use you \cr
   \code{aqipxc037} \tab 0/1 \tab Show your child how to make a bridge with blocks, boxes, or \cr
   \code{aqipxc038} \tab 0/1 \tab When you say, "Say seven three," does your child repeat just \cr
   \code{aqipxc039} \tab 0/1 \tab Does your child dress up and "play-act," pretending to be so \cr
   \code{aqipxc040} \tab 0/1 \tab When you point to the figure and ask your child, "What is th \cr
   \code{aqipxc042} \tab 0/1 \tab When you say, "Say five eight three," does your child repeat \cr
   \code{aqipxc043} \tab 0/1 \tab After your child draws a "picture," even a simple scribble, \cr
   \code{aqipxc045} \tab 0/1 \tab When asked, "Which circle is the smallest?" does your child \cr
   \code{aqislc010} \tab 0/1 \tab While lying on her back, does your baby play by grabbing her \cr
   \code{aqislc011} \tab 0/1 \tab Does your baby try to get a toy that is out of reach? (She m \cr
   \code{aqislc013} \tab 0/1 \tab When in front of a large mirror, does your baby reach out to \cr
   \code{aqislc014} \tab 0/1 \tab Does your baby feed himself a cracker or a cookie? \cr
   \code{aqislc015} \tab 0/1 \tab Does your baby drink water, juice, or formula from a cup whi \cr
   \code{aqislc017} \tab 0/1 \tab While your baby is on his back, does he put his foot in his \cr
   \code{aqislc018} \tab 0/1 \tab When you dress your baby, does he push his arm through a sle \cr
   \code{aqislc019} \tab 0/1 \tab When you hold out your hand and ask for her toy, does your b \cr
   \code{aqislc020} \tab 0/1 \tab When you hold out your hand and ask for her toy, does your b \cr
   \code{aqislc021} \tab 0/1 \tab Does your baby roll or throw a ball back to you so that you \cr
   \code{aqislc022} \tab 0/1 \tab Does your baby/child play with a doll or stuffed animal by h \cr
   \code{aqislc023} \tab 0/1 \tab When you dress your baby does she lift her foot for her shoe \cr
   \code{aqislc024} \tab 0/1 \tab Does your child copy the activities you do, such as wipe up \cr
   \code{aqislc025} \tab 0/1 \tab Does your child push a little wagon, stroller or other toy o \cr
   \code{aqislc026} \tab 0/1 \tab Does your baby/child feed herself with a spoon, even though \cr
   \code{aqislc027} \tab 0/1 \tab Does your baby/child get your attention or try to show you s \cr
   \code{aqislc028} \tab 0/1 \tab Does your child come to you when she needs help, such as wit \cr
   \code{aqislc029} \tab 0/1 \tab Does your child use a spoon to feed himself with little spil \cr
   \code{aqislc030} \tab 0/1 \tab Does your child drink from a cup or glass, putting it down a \cr
   \code{aqislc031} \tab 0/1 \tab If you do any of the following gestures, does your child cop \cr
   \code{aqislc032} \tab 0/1 \tab Does your baby/child help undress himself by taking off clot \cr
   \code{aqislc033} \tab 0/1 \tab Does your child eat with a fork? \cr
   \code{aqislc034} \tab 0/1 \tab Does your child take turns by waiting while another child or \cr
   \code{aqislc035} \tab 0/1 \tab When playing with either a stuffed animal or doll, does your \cr
   \code{aqislc036} \tab 0/1 \tab When your child is looking in a mirror and you ask, "Who is \cr
   \code{aqislc037} \tab 0/1 \tab While looking at himself in the mirror, does your child offe \cr
   \code{aqislc038} \tab 0/1 \tab After you put on loose-fitting pants around her feet, does y \cr
   \code{aqislc039} \tab 0/1 \tab Does your child call herself "I" or "me" more often than her \cr
   \code{aqislc041} \tab 0/1 \tab Using these exact words, ask your child, "Are you a girl or \cr
   \code{aqislc048} \tab 0/1 \tab Does your child put on a coat, jacket, or shirt by himself? \cr
   \code{aqislc053} \tab 0/1 \tab Does your child serve herself, taking food from one containe \cr
   \code{batadd001} \tab 0/1 \tab Takes strained food from spoon and swallows it (TS21) \cr
   \code{batadd003} \tab 0/1 \tab Place both hands on a bottle or breast during feeding (TS23) \cr
   \code{batadd004} \tab 0/1 \tab Eats semisolid food when it is placed in his or her mouth (T \cr
   \code{batadd005} \tab 0/1 \tab Begins to use spoon or other utensil to feed self (TS25) \cr
   \code{batadd011} \tab 0/1 \tab Sucks with smooth, coordinated movements \cr
   \code{batadd012} \tab 0/1 \tab Helps dress self by holding out his/her arms or legs \cr
   \code{batadd013} \tab 0/1 \tab Asks for foor or liquid with words or gestures \cr
   \code{batadd014} \tab 0/1 \tab Removes shoes without assistance \cr
   \code{batadd015} \tab 0/1 \tab Feeds self with a spoon or fork without assistance \cr
   \code{batadd016} \tab 0/1 \tab Puts away toys when asked \cr
   \code{batadd017} \tab 0/1 \tab Blows his/her nose with assistance \cr
   \code{batadd018} \tab 0/1 \tab Washes and dries hands without assistance \cr
   \code{batadd019} \tab 0/1 \tab Chooses the appropriate ustensil for the food he/she is eati \cr
   \code{batadd020} \tab 0/1 \tab Responds to instructions given in a small group and initiate \cr
   \code{batadd021} \tab 0/1 \tab Cuts soft foods with the side of a fork \cr
   \code{batadd022} \tab 0/1 \tab Answers "what-to-do-if" questions involving personal respons \cr
   \code{batadd023} \tab 0/1 \tab Chooses clothing that is approprate for the weather \cr
   \code{batadd024} \tab 0/1 \tab Knows his/her own phone number \cr
   \code{batadd025} \tab 0/1 \tab Goes to bed without assistance \cr
   \code{batadd026} \tab 0/1 \tab Uses emergency phone numbers \cr
   \code{batcgd002} \tab 0/1 \tab Feels and explore objects (TS 8) \cr
   \code{batcgd003} \tab 0/1 \tab Uncovers hidden toy (TS 81) \cr
   \code{batcgd009} \tab 0/1 \tab Visually attends to a light source moving in a 180-degree ar \cr
   \code{batcgd010} \tab 0/1 \tab Turns eyes towards a light source \cr
   \code{batcgd011} \tab 0/1 \tab Attends to an ongoing sound or activity for 15 or more secon \cr
   \code{batcgd012} \tab 0/1 \tab Attends to a game of pekaboo for 1 min \cr
   \code{batcgd013} \tab 0/1 \tab Looks at, points to, or touches pictures in a book \cr
   \code{batcgd014} \tab 0/1 \tab Imitates simple facial gestures \cr
   \code{batcgd015} \tab 0/1 \tab Finds an object hidden under one of two cups \cr
   \code{batcgd016} \tab 0/1 \tab Nests objects inside one another \cr
   \code{batcgd017} \tab 0/1 \tab Locates hidden items in a picture scene - Level 1 \cr
   \code{batcgd018} \tab 0/1 \tab Names the colores red, green, and blue \cr
   \code{batcgd019} \tab 0/1 \tab Locates hidden items in a picture scene - Level 2 \cr
   \code{batcgd020} \tab 0/1 \tab Recognises visual differences among similar numerals and let \cr
   \code{batcgd021} \tab 0/1 \tab Identifies the picture that is different \cr
   \code{batcgd022} \tab 0/1 \tab Categorises familiar objects by function \cr
   \code{batcgd023} \tab 0/1 \tab Matches simple words \cr
   \code{batcgd024} \tab 0/1 \tab Knows the right and left sides of his/her body \cr
   \code{batcgd025} \tab 0/1 \tab Repeats sequences of 4 and 5 objects from memory \cr
   \code{batcgd026} \tab 0/1 \tab Groups objects by shape and color \cr
   \code{batcmd004} \tab 0/1 \tab Produces one or more single-syllable consonant-vowel sounds \cr
   \code{batcmd009} \tab 0/1 \tab Is soothed by a familiar adult's voice \cr
   \code{batcmd010} \tab 0/1 \tab Produces differenciated cries \cr
   \code{batcmd011} \tab 0/1 \tab Responds to different tones of a person's voice \cr
   \code{batcmd012} \tab 0/1 \tab Identifies family members or pets when named \cr
   \code{batcmd013} \tab 0/1 \tab Uses variations in his/her voice \cr
   \code{batcmd014} \tab 0/1 \tab Follows 3 or more familiar verbal commands \cr
   \code{batcmd015} \tab 0/1 \tab Spontaneously initiates sounds, words, or gesturres that are \cr
   \code{batcmd016} \tab 0/1 \tab Responds to the prepositions out and on \cr
   \code{batcmd017} \tab 0/1 \tab Uses 2-words utterances to express meaningful relationships \cr
   \code{batcmd018} \tab 0/1 \tab Responds to who and what questions \cr
   \code{batcmd019} \tab 0/1 \tab Uses words to relate information about other people, their a \cr
   \code{batcmd020} \tab 0/1 \tab Responds to where and when questions \cr
   \code{batcmd021} \tab 0/1 \tab Repeats familiar words with clear articulation \cr
   \code{batcmd022} \tab 0/1 \tab Converses on topics for more than 5 turn-taking exchanges \cr
   \code{batcmd023} \tab 0/1 \tab Identifies a word from its definition \cr
   \code{batcmd024} \tab 0/1 \tab Follows 3-step verbal commands \cr
   \code{batcmd025} \tab 0/1 \tab Uses plural forms ending in the /ez/ sound \cr
   \code{batcmd026} \tab 0/1 \tab Recalls events from a sotry presented orally \cr
   \code{batcmd027} \tab 0/1 \tab Describes what is happening in a picture \cr
   \code{batmod011} \tab 0/1 \tab Mantains an upright posture at adult's shoulder without assi \cr
   \code{batmod012} \tab 0/1 \tab Holds hands in an open, loose-fisted position when not grasp \cr
   \code{batmod013} \tab 0/1 \tab Retrieves a small object by raking it with his/her fingers a \cr
   \code{batmod014} \tab 0/1 \tab Transfers an object from one hand to the other \cr
   \code{batmod015} \tab 0/1 \tab Moves from a sitting position to a standing position without \cr
   \code{batmod016} \tab 0/1 \tab Intentionally propels or throws an object \cr
   \code{batmod017} \tab 0/1 \tab Maintains or corrects his/her balance when moving from a sta \cr
   \code{batmod018} \tab 0/1 \tab Removes forms from a form board \cr
   \code{batmod019} \tab 0/1 \tab Runs 1 feet without falling \cr
   \code{batmod020} \tab 0/1 \tab Scribbles linear and/or circular patterns spontaneously \cr
   \code{batmod021} \tab 0/1 \tab Walks forward 2 or more steps on a line on the floor, altern \cr
   \code{batmod022} \tab 0/1 \tab Stacks 8 cubes vertically \cr
   \code{batmod023} \tab 0/1 \tab Hops forward on one foot without support \cr
   \code{batmod024} \tab 0/1 \tab Folds a sheet of paper \cr
   \code{batmod025} \tab 0/1 \tab Touches the fingertips of each hand successively with the th \cr
   \code{batmod026} \tab 0/1 \tab Draws a person with 6 parts \cr
   \code{batmod027} \tab 0/1 \tab Walks a 6-foot line on the floor, heel-to-toe, with eyes ope \cr
   \code{batmod028} \tab 0/1 \tab Copies the numerals 1 through 5 \cr
   \code{batmod029} \tab 0/1 \tab Skips on alternate feet for 20 feet \cr
   \code{batmod030} \tab 0/1 \tab Ties a single overhand knot around a crayon with a string \cr
   \code{batsld005} \tab 0/1 \tab Initiates social contact with peers in play (TS5) \cr
   \code{batsld009} \tab 0/1 \tab Knows his/her first name (TS9) \cr
   \code{batsld011} \tab 0/1 \tab Shows awareness of other people \cr
   \code{batsld012} \tab 0/1 \tab Smiles or vocalizes in response to adult attention \cr
   \code{batsld013} \tab 0/1 \tab Shows a desire for social attention \cr
   \code{batsld014} \tab 0/1 \tab Is aware of his/her feet \cr
   \code{batsld015} \tab 0/1 \tab Discriminates between familiar and unfamiliar people \cr
   \code{batsld016} \tab 0/1 \tab Displays independent behavior \cr
   \code{batsld017} \tab 0/1 \tab Greets familiar adults spontaneously \cr
   \code{batsld018} \tab 0/1 \tab Responds positively when familiar adults or adults in author \cr
   \code{batsld019} \tab 0/1 \tab Responds differently to familiar and unfamiliar children \cr
   \code{batsld020} \tab 0/1 \tab Allow others to participate in his/her activities \cr
   \code{batsld021} \tab 0/1 \tab Engages in adult role-playing and imitation \cr
   \code{batsld022} \tab 0/1 \tab Follows adult directions with little or no resistance \cr
   \code{batsld023} \tab 0/1 \tab Recognizes an adult's feelings \cr
   \code{batsld024} \tab 0/1 \tab Cooperates in group activities \cr
   \code{batsld025} \tab 0/1 \tab Discriminates between socially acceptable and unacceptable b \cr
   \code{batsld026} \tab 0/1 \tab Trusts familiar adults and accepts explanations from them \cr
   \code{batsld027} \tab 0/1 \tab Waits his/her turn for a teacher's or adult's attention \cr
   \code{batsld028} \tab 0/1 \tab Delays gratification until a task is completed \cr
   \code{by3cgd022} \tab 0/1 \tab Mirror Image Series: Responds positively \cr
   \code{by3cgd023} \tab 0/1 \tab Plays with string \cr
   \code{by3cgd024} \tab 0/1 \tab Bangs in play \cr
   \code{by3cgd025} \tab 0/1 \tab Searches for fallen object \cr
   \code{by3cgd026} \tab 0/1 \tab Bell Series: Manipulates \cr
   \code{by3cgd027} \tab 0/1 \tab Picks Up Block Series: Reaches for second block \cr
   \code{by3cgd028} \tab 0/1 \tab Pulls cloth to obtain object \cr
   \code{by3cgd029} \tab 0/1 \tab Pulls string adaptively \cr
   \code{by3cgd030} \tab 0/1 \tab Retains both blocks \cr
   \code{by3cgd031} \tab 0/1 \tab Bell Series: Rings purposely \cr
   \code{by3cgd032} \tab 0/1 \tab Looks at pictures \cr
   \code{by3cgd033} \tab 0/1 \tab Picks Up Block Series: Retains 2 of 3 blocks \cr
   \code{by3cgd034} \tab 0/1 \tab Searches for missing objects \cr
   \code{by3cgd035} \tab 0/1 \tab Takes blocks out of cup \cr
   \code{by3cgd036} \tab 0/1 \tab Block Series: 1 Block \cr
   \code{by3cgd037} \tab 0/1 \tab Picks up Block Series: 3 blocks \cr
   \code{by3cgd038} \tab 0/1 \tab Explores holes in pegboard \cr
   \code{by3cgd039} \tab 0/1 \tab Pushes car \cr
   \code{by3cgd040} \tab 0/1 \tab Finds hidden object \cr
   \code{by3cgd041} \tab 0/1 \tab Suspends ring \cr
   \code{by3cgd042} \tab 0/1 \tab Removes pellet \cr
   \code{by3cgd043} \tab 0/1 \tab Clear Box: Front \cr
   \code{by3cgd044} \tab 0/1 \tab Squeezes object \cr
   \code{by3cgd045} \tab 0/1 \tab Finds hidden object (Reversed) \cr
   \code{by3cgd046} \tab 0/1 \tab Removes lid from bottle \cr
   \code{by3cgd047} \tab 0/1 \tab Pegboard Series: 2 holes \cr
   \code{by3cgd048} \tab 0/1 \tab Relational Play Series: Self \cr
   \code{by3cgd049} \tab 0/1 \tab Pink Board Series: 1 piece \cr
   \code{by3cgd050} \tab 0/1 \tab Finds hidden object (Visible Displacement) \cr
   \code{by3cgd051} \tab 0/1 \tab Blue Board Series: 1 piece \cr
   \code{by3cgd052} \tab 0/1 \tab Clear Box: Sides \cr
   \code{by3cgd053} \tab 0/1 \tab Relational Play Series: Others \cr
   \code{by3cgd054} \tab 0/1 \tab Block Series: 9 Blocks \cr
   \code{by3cgd055} \tab 0/1 \tab Pegboard Series: 6 Pegs \cr
   \code{by3cgd056} \tab 0/1 \tab Pink Board Series: Completes \cr
   \code{by3cgd057} \tab 0/1 \tab Uses pencil to obtain object \cr
   \code{by3cgd058} \tab 0/1 \tab Blue Board Series: 4 Pieces \cr
   \code{by3cgd059} \tab 0/1 \tab Attends to story \cr
   \code{by3cgd060} \tab 0/1 \tab Rotated pink board \cr
   \code{by3cgd061} \tab 0/1 \tab Object assembly (Ball) \cr
   \code{by3cgd062} \tab 0/1 \tab Completes Pegboard: 25 Seconds \cr
   \code{by3cgd063} \tab 0/1 \tab Object assembly (Ice Cream Cone) \cr
   \code{by3cgd064} \tab 0/1 \tab Matches pictures \cr
   \code{by3cgd065} \tab 0/1 \tab Representational play \cr
   \code{by3cgd066} \tab 0/1 \tab Blue Board Series: Completes (75 seconds) \cr
   \code{by3cgd067} \tab 0/1 \tab Imitates a two-step action. \cr
   \code{by3cgd068} \tab 0/1 \tab Matches 3 colors \cr
   \code{by3cgd069} \tab 0/1 \tab Imaginary play \cr
   \code{by3cgd070} \tab 0/1 \tab Understands concept of one \cr
   \code{by3cgd071} \tab 0/1 \tab Multischeme combination play \cr
   \code{by3cgd072} \tab 0/1 \tab Concept Grouping: colour \cr
   \code{by3cgd073} \tab 0/1 \tab Concept Grouping: Size \cr
   \code{by3cgd074} \tab 0/1 \tab Compares masses \cr
   \code{by3cgd075} \tab 0/1 \tab Matches size \cr
   \code{by3cgd076} \tab 0/1 \tab Discriminates pictures \cr
   \code{by3cgd077} \tab 0/1 \tab Simple pattern \cr
   \code{by3cgd078} \tab 0/1 \tab Sorts pegs by colour \cr
   \code{by3cgd079} \tab 0/1 \tab Counts (One-to-one correspondence) \cr
   \code{by3cgd080} \tab 0/1 \tab Discriminates size \cr
   \code{by3cgd081} \tab 0/1 \tab Identifies 3 incomplete pictures \cr
   \code{by3cgd082} \tab 0/1 \tab Object assembly (Dog) \cr
   \code{by3cgd083} \tab 0/1 \tab Discriminates patterns \cr
   \code{by3cgd084} \tab 0/1 \tab Spatial memory \cr
   \code{by3cgd085} \tab 0/1 \tab Counts (Cardinally) \cr
   \code{by3cgd086} \tab 0/1 \tab Number constancy \cr
   \code{by3cgd087} \tab 0/1 \tab Laces card \cr
   \code{by3cgd088} \tab 0/1 \tab Classifies objects \cr
   \code{by3cgd089} \tab 0/1 \tab Understands concept of more \cr
   \code{by3exd003} \tab 0/1 \tab Vocalizes mood \cr
   \code{by3exd005} \tab 0/1 \tab Social vocalization or laughing \cr
   \code{by3exd006} \tab 0/1 \tab 2 Vowel sounds \cr
   \code{by3exd007} \tab 0/1 \tab Gets attention \cr
   \code{by3exd008} \tab 0/1 \tab 2 Consonant sounds \cr
   \code{by3exd009} \tab 0/1 \tab Uses gestures \cr
   \code{by3exd010} \tab 0/1 \tab Consonant-Vowel Combination Series: 1 Combination \cr
   \code{by3exd011} \tab 0/1 \tab Participates in play routine \cr
   \code{by3exd012} \tab 0/1 \tab Jabbers expressively \cr
   \code{by3exd013} \tab 0/1 \tab Consonant-Vowel Combination Series: 4 Combination \cr
   \code{by3exd014} \tab 0/1 \tab Uses one-word approximations \cr
   \code{by3exd015} \tab 0/1 \tab Directs attention of other \cr
   \code{by3exd016} \tab 0/1 \tab Imitates word \cr
   \code{by3exd017} \tab 0/1 \tab Initiates play interaction \cr
   \code{by3exd018} \tab 0/1 \tab Uses Words Appropriately Series: 2 words \cr
   \code{by3exd019} \tab 0/1 \tab Uses word to make wants known \cr
   \code{by3exd020} \tab 0/1 \tab Names Object Series: 1 Object \cr
   \code{by3exd021} \tab 0/1 \tab Combines word and gesture \cr
   \code{by3exd022} \tab 0/1 \tab Names Picture Series: 1 Picture \cr
   \code{by3exd023} \tab 0/1 \tab Uses Words Appropriately Series: 8 words \cr
   \code{by3exd024} \tab 0/1 \tab Answers yes or no verbally in response to questions \cr
   \code{by3exd025} \tab 0/1 \tab Imitates a two-word utterance \cr
   \code{by3exd026} \tab 0/1 \tab Uses a two-word utterance \cr
   \code{by3exd027} \tab 0/1 \tab Names Object Series: 3 Object \cr
   \code{by3exd028} \tab 0/1 \tab Names Picture Series: 5 Pictures \cr
   \code{by3exd029} \tab 0/1 \tab Uses multiple-word utterances \cr
   \code{by3exd030} \tab 0/1 \tab Uses pronouns \cr
   \code{by3exd031} \tab 0/1 \tab Names Action Picture Series: 1 Picture \cr
   \code{by3exd032} \tab 0/1 \tab Poses multiple-word questions \cr
   \code{by3exd033} \tab 0/1 \tab Makes a contingent utterance \cr
   \code{by3exd034} \tab 0/1 \tab Uses verb +ing \cr
   \code{by3exd035} \tab 0/1 \tab Names Action Picture Series: 3 Pictures \cr
   \code{by3exd036} \tab 0/1 \tab Uses different word combinations \cr
   \code{by3exd037} \tab 0/1 \tab Names Action Picture Series: 5 Pictures \cr
   \code{by3exd038} \tab 0/1 \tab Uses plurals \cr
   \code{by3exd039} \tab 0/1 \tab Answers what or where questions \cr
   \code{by3exd040} \tab 0/1 \tab Uses possessives \cr
   \code{by3exd041} \tab 0/1 \tab Names 4 colours \cr
   \code{by3exd042} \tab 0/1 \tab Answers questions logically (Related to functions) \cr
   \code{by3exd043} \tab 0/1 \tab Tells how an object is used \cr
   \code{by3exd044} \tab 0/1 \tab Uses prepositions \cr
   \code{by3exd045} \tab 0/1 \tab Uses present progressive form \cr
   \code{by3exd046} \tab 0/1 \tab Describes Pictures Series: Uses 4-5 word sentences \cr
   \code{by3exd047} \tab 0/1 \tab Describes Pictures Series: Uses past tense \cr
   \code{by3exd048} \tab 0/1 \tab Describes Pictures Series: Uses future tense \cr
   \code{by3fmd016} \tab 0/1 \tab Reaches unilaterally \cr
   \code{by3fmd017} \tab 0/1 \tab Food Pellet Series: Raking grasp \cr
   \code{by3fmd018} \tab 0/1 \tab Block Series: Partial thumb opposition \cr
   \code{by3fmd019} \tab 0/1 \tab Transfers ring \cr
   \code{by3fmd020} \tab 0/1 \tab Food Pellet Series: Whole hand grasp \cr
   \code{by3fmd021} \tab 0/1 \tab Transfers block \cr
   \code{by3fmd022} \tab 0/1 \tab Block Series: Thumb-fingertip grasp \cr
   \code{by3fmd023} \tab 0/1 \tab Brings spoons or blocks to midline \cr
   \code{by3fmd024} \tab 0/1 \tab Food Pellet Series: Partial thumb opposition \cr
   \code{by3fmd025} \tab 0/1 \tab Lifts cup by the handle \cr
   \code{by3fmd026} \tab 0/1 \tab Food Pellet Series: Thumb-fingertip grasp \cr
   \code{by3fmd027} \tab 0/1 \tab Turns pages of books \cr
   \code{by3fmd028} \tab 0/1 \tab Grasp series: Palmar grasp \cr
   \code{by3fmd029} \tab 0/1 \tab Isolates extended ring finger \cr
   \code{by3fmd030} \tab 0/1 \tab Scribbles spontaneously \cr
   \code{by3fmd031} \tab 0/1 \tab Block Stacking Series: 2 blocks \cr
   \code{by3fmd032} \tab 0/1 \tab Imitates Stroke Series: Random \cr
   \code{by3fmd033} \tab 0/1 \tab Places 10 pellets in bottle (60 seconds) \cr
   \code{by3fmd034} \tab 0/1 \tab Grasp series: Transitional grasp \cr
   \code{by3fmd035} \tab 0/1 \tab Coins in slot \cr
   \code{by3fmd036} \tab 0/1 \tab Connecting Blocks: Apart \cr
   \code{by3fmd037} \tab 0/1 \tab Grasp Series: Intermediate (Tripod) grasp \cr
   \code{by3fmd038} \tab 0/1 \tab Block stacking Series: 6 blocks \cr
   \code{by3fmd039} \tab 0/1 \tab Uses hand to hold paper in place \cr
   \code{by3fmd040} \tab 0/1 \tab Imitates Strokes Series: Horizontal \cr
   \code{by3fmd041} \tab 0/1 \tab Imitates Strokes Series: Vertical \cr
   \code{by3fmd042} \tab 0/1 \tab Connecting Blocks: Together \cr
   \code{by3fmd043} \tab 0/1 \tab Imitates Strokes Series: Circular \cr
   \code{by3fmd044} \tab 0/1 \tab Builds train of blocks \cr
   \code{by3fmd045} \tab 0/1 \tab Strings 3 blocks \cr
   \code{by3fmd046} \tab 0/1 \tab Imitates hand movements \cr
   \code{by3fmd047} \tab 0/1 \tab Snips paper \cr
   \code{by3fmd048} \tab 0/1 \tab Grasp Series: Dynamic grasp \cr
   \code{by3fmd049} \tab 0/1 \tab Tactilely discriminates shapes \cr
   \code{by3fmd050} \tab 0/1 \tab Builds wall \cr
   \code{by3fmd051} \tab 0/1 \tab Cuts paper \cr
   \code{by3fmd052} \tab 0/1 \tab Builds bridge \cr
   \code{by3fmd053} \tab 0/1 \tab Imitates plus sign \cr
   \code{by3fmd054} \tab 0/1 \tab Block stacking Series: 8 blocks \cr
   \code{by3fmd055} \tab 0/1 \tab Cuts on line \cr
   \code{by3fmd056} \tab 0/1 \tab Builds T \cr
   \code{by3fmd057} \tab 0/1 \tab Buttons 1 button \cr
   \code{by3fmd058} \tab 0/1 \tab Builds steps \cr
   \code{by3fmd059} \tab 0/1 \tab Traces designs \cr
   \code{by3fmd060} \tab 0/1 \tab Imitates square \cr
   \code{by3fmd061} \tab 0/1 \tab Copies plus sign \cr
   \code{by3fmd062} \tab 0/1 \tab Tapbs fingers \cr
   \code{by3fmd063} \tab 0/1 \tab Places 20 pellets in bottle \cr
   \code{by3fmd064} \tab 0/1 \tab Cuts circle \cr
   \code{by3fmd065} \tab 0/1 \tab Cuts square \cr
   \code{by3fmd066} \tab 0/1 \tab Copies square \cr
   \code{by3gmd018} \tab 0/1 \tab Elevates trunk while prone \cr
   \code{by3gmd019} \tab 0/1 \tab Sits with Support Series: 30 seconds \cr
   \code{by3gmd020} \tab 0/1 \tab Rolls from back to sides \cr
   \code{by3gmd021} \tab 0/1 \tab Elevates Trunk while Prone Series: Extended arms \cr
   \code{by3gmd022} \tab 0/1 \tab Sits with Support Series: 5 seconds \cr
   \code{by3gmd023} \tab 0/1 \tab Pulls up to sit \cr
   \code{by3gmd024} \tab 0/1 \tab Grasps foot with hands \cr
   \code{by3gmd025} \tab 0/1 \tab Rolls from back to stomach \cr
   \code{by3gmd026} \tab 0/1 \tab Sits without Support Series: 30 seconds \cr
   \code{by3gmd027} \tab 0/1 \tab Sits without support and holds objects \cr
   \code{by3gmd028} \tab 0/1 \tab Rotates trunk while seated \cr
   \code{by3gmd029} \tab 0/1 \tab Makes stepping movements \cr
   \code{by3gmd030} \tab 0/1 \tab Crawls Series: On stomach \cr
   \code{by3gmd031} \tab 0/1 \tab Crawls Series: Crawl position \cr
   \code{by3gmd032} \tab 0/1 \tab Moves from sitting to hands and knees \cr
   \code{by3gmd033} \tab 0/1 \tab Supports weight \cr
   \code{by3gmd034} \tab 0/1 \tab Crawls Series: Crawl movement \cr
   \code{by3gmd035} \tab 0/1 \tab Raises self to standing position \cr
   \code{by3gmd036} \tab 0/1 \tab Bounces while standing \cr
   \code{by3gmd037} \tab 0/1 \tab Walks Series: With support \cr
   \code{by3gmd038} \tab 0/1 \tab Walks sideways with support \cr
   \code{by3gmd039} \tab 0/1 \tab Sits down with control \cr
   \code{by3gmd040} \tab 0/1 \tab Stands alone \cr
   \code{by3gmd041} \tab 0/1 \tab Stands up Series: Alone \cr
   \code{by3gmd042} \tab 0/1 \tab Walks Series: Alone \cr
   \code{by3gmd043} \tab 0/1 \tab Walks Series: Alone with coordination \cr
   \code{by3gmd044} \tab 0/1 \tab Throws ball \cr
   \code{by3gmd045} \tab 0/1 \tab Squats without support \cr
   \code{by3gmd046} \tab 0/1 \tab Stands up Series: Mature \cr
   \code{by3gmd047} \tab 0/1 \tab Walks Up Stairs Series: Both feet on each step, with support \cr
   \code{by3gmd048} \tab 0/1 \tab Walks backward 2 steps \cr
   \code{by3gmd049} \tab 0/1 \tab Walks Down Stairs Series: Both feet on each step, with suppo \cr
   \code{by3gmd050} \tab 0/1 \tab Runs with coordination \cr
   \code{by3gmd051} \tab 0/1 \tab Balances on Right Foot Series: With support \cr
   \code{by3gmd052} \tab 0/1 \tab Balances on Left Foot Series: With support \cr
   \code{by3gmd053} \tab 0/1 \tab Walks sideways without support \cr
   \code{by3gmd054} \tab 0/1 \tab Jumps from bottom step \cr
   \code{by3gmd055} \tab 0/1 \tab Kicks ball \cr
   \code{by3gmd056} \tab 0/1 \tab Walks forward on path \cr
   \code{by3gmd057} \tab 0/1 \tab Walks Up Stairs Series: Both feet on each step, alone. \cr
   \code{by3gmd058} \tab 0/1 \tab Walks Down Stairs Series: Both feet on each step, alone \cr
   \code{by3gmd059} \tab 0/1 \tab Jumps Forward Series: 4 inches \cr
   \code{by3gmd060} \tab 0/1 \tab Balances on right Foot Series: 2 seconds, alone \cr
   \code{by3gmd061} \tab 0/1 \tab Balances on Left Foot Series: 2 seconds, alone \cr
   \code{by3gmd062} \tab 0/1 \tab Walks on tiptoes 4 steps \cr
   \code{by3gmd063} \tab 0/1 \tab Walks backward close to path \cr
   \code{by3gmd064} \tab 0/1 \tab Walks Up Stairs Series: Alternating feet, alone \cr
   \code{by3gmd065} \tab 0/1 \tab Imitates postures \cr
   \code{by3gmd066} \tab 0/1 \tab Stops from a full run \cr
   \code{by3gmd067} \tab 0/1 \tab Walks Down Stairs Series: Alternating feet, alone \cr
   \code{by3gmd068} \tab 0/1 \tab Hops 5 feet \cr
   \code{by3gmd069} \tab 0/1 \tab Balances on Right Foot Series: 8 seconds, alone \cr
   \code{by3gmd070} \tab 0/1 \tab Balances on Left Foot Series: 8 seconds, alone \cr
   \code{by3gmd071} \tab 0/1 \tab Walks heel to toe \cr
   \code{by3gmd072} \tab 0/1 \tab Jumps Forward Series: 24 inches \cr
   \code{by3red006} \tab 0/1 \tab Searches with head turn \cr
   \code{by3red007} \tab 0/1 \tab Discriminates sounds \cr
   \code{by3red008} \tab 0/1 \tab Sustained play with objects \cr
   \code{by3red009} \tab 0/1 \tab Responds to name \cr
   \code{by3red010} \tab 0/1 \tab Interrupts activity \cr
   \code{by3red011} \tab 0/1 \tab Recognizes 2 familiar words \cr
   \code{by3red012} \tab 0/1 \tab Responds to no-no \cr
   \code{by3red013} \tab 0/1 \tab Attends to other's play routine \cr
   \code{by3red014} \tab 0/1 \tab Responds to request for social routines \cr
   \code{by3red015} \tab 0/1 \tab Identifies Object Series: 1 correct \cr
   \code{by3red016} \tab 0/1 \tab Identifies object in the environment \cr
   \code{by3red017} \tab 0/1 \tab Identifies Picture Series: 1 correct \cr
   \code{by3red018} \tab 0/1 \tab Understands inhibitory words \cr
   \code{by3red019} \tab 0/1 \tab Identifies Object Series: 3 correct \cr
   \code{by3red020} \tab 0/1 \tab Follows one-part directions \cr
   \code{by3red021} \tab 0/1 \tab Identifies Picture Series: 3 Correct \cr
   \code{by3red022} \tab 0/1 \tab Identifies 3 clothing items \cr
   \code{by3red023} \tab 0/1 \tab Identifies Action Picture Series: 1 correct \cr
   \code{by3red024} \tab 0/1 \tab Identifies 5 parts of the body \cr
   \code{by3red025} \tab 0/1 \tab Follows two-part directions \cr
   \code{by3red026} \tab 0/1 \tab Identifies Action Picture Series: 3 correct \cr
   \code{by3red027} \tab 0/1 \tab Understands use of objects \cr
   \code{by3red028} \tab 0/1 \tab Understands part/whole relationships \cr
   \code{by3red029} \tab 0/1 \tab Identifies Action Picture Series: 5 correct \cr
   \code{by3red030} \tab 0/1 \tab Understands pronouns (him, me, my, you, your) \cr
   \code{by3red031} \tab 0/1 \tab Understands labels for sizes \cr
   \code{by3red032} \tab 0/1 \tab Understands Preposition Series: 2 correct \cr
   \code{by3red033} \tab 0/1 \tab Understands possessives \cr
   \code{by3red034} \tab 0/1 \tab Understands verb +ing \cr
   \code{by3red035} \tab 0/1 \tab Identifies colours \cr
   \code{by3red036} \tab 0/1 \tab Understands label of one \cr
   \code{by3red037} \tab 0/1 \tab Understands pronouns (they, he, she) \cr
   \code{by3red038} \tab 0/1 \tab Understands pronouns (his, her) \cr
   \code{by3red039} \tab 0/1 \tab Understands plurals \cr
   \code{by3red040} \tab 0/1 \tab Understands more \cr
   \code{by3red041} \tab 0/1 \tab Understands most \cr
   \code{by3red042} \tab 0/1 \tab Understands Preposition Series: 4 correct \cr
   \code{by3red043} \tab 0/1 \tab Understands negatives in sentences \cr
   \code{by3red044} \tab 0/1 \tab Understands past tense \cr
   \code{by3red045} \tab 0/1 \tab Understands labels for mass \cr
   \code{denfmd003} \tab 0/1 \tab Hold Rattle \cr
   \code{denfmd004} \tab 0/1 \tab Hands Together \cr
   \code{denfmd005} \tab 0/1 \tab Follow 180 Degrees \cr
   \code{denfmd006} \tab 0/1 \tab Regard raisin \cr
   \code{denfmd007} \tab 0/1 \tab Reaches \cr
   \code{denfmd008} \tab 0/1 \tab Look for Yarn \cr
   \code{denfmd009} \tab 0/1 \tab Rake Raisin \cr
   \code{denfmd010} \tab 0/1 \tab Pass Block \cr
   \code{denfmd011} \tab 0/1 \tab Take 2 Blocks \cr
   \code{denfmd012} \tab 0/1 \tab Thumb-Finger Grip \cr
   \code{denfmd013} \tab 0/1 \tab Bang 2 Blocks \cr
   \code{denfmd014} \tab 0/1 \tab Put Block in Cup \cr
   \code{denfmd015} \tab 0/1 \tab Scribbles \cr
   \code{denfmd016} \tab 0/1 \tab Drop Raisin \cr
   \code{denfmd017} \tab 0/1 \tab 2 Blocks \cr
   \code{denfmd018} \tab 0/1 \tab 4 Blocks \cr
   \code{denfmd019} \tab 0/1 \tab 6 Blocks \cr
   \code{denfmd020} \tab 0/1 \tab Imitate Vertical Line \cr
   \code{denfmd021} \tab 0/1 \tab 8 Blocks \cr
   \code{denfmd022} \tab 0/1 \tab Wiggle Thumb \cr
   \code{denfmd023} \tab 0/1 \tab Copy a Circle \cr
   \code{denfmd024} \tab 0/1 \tab Draws 3 or more body parts \cr
   \code{denfmd025} \tab 0/1 \tab Copy +  (2 lines that intersect near the midpoint using only \cr
   \code{denfmd026} \tab 0/1 \tab Pick Longer Line \cr
   \code{denfmd027} \tab 0/1 \tab Copy Square, Demonstrated \cr
   \code{denfmd028} \tab 0/1 \tab Draws 6 or more body parts \cr
   \code{denfmd029} \tab 0/1 \tab Copy Square \cr
   \code{dengmd007} \tab 0/1 \tab Lift Chest with Arms \cr
   \code{dengmd008} \tab 0/1 \tab Roll Over/from back to side \cr
   \code{dengmd009} \tab 0/1 \tab Pull to Sit, No Head Lag \cr
   \code{dengmd010} \tab 0/1 \tab Sit No Support \cr
   \code{dengmd011} \tab 0/1 \tab Stand, Holding On \cr
   \code{dengmd012} \tab 0/1 \tab Pull to Stand \cr
   \code{dengmd013} \tab 0/1 \tab Get To Sitting \cr
   \code{dengmd014} \tab 0/1 \tab Stand 2 Seconds \cr
   \code{dengmd015} \tab 0/1 \tab Stand 10 seconds \cr
   \code{dengmd016} \tab 0/1 \tab Stoop and Recover \cr
   \code{dengmd017} \tab 0/1 \tab Walk Well \cr
   \code{dengmd018} \tab 0/1 \tab Walk Backwards \cr
   \code{dengmd019} \tab 0/1 \tab Runs \cr
   \code{dengmd020} \tab 0/1 \tab Walk Up Stairs \cr
   \code{dengmd021} \tab 0/1 \tab Kick Ball Forward \cr
   \code{dengmd022} \tab 0/1 \tab Jump Up \cr
   \code{dengmd023} \tab 0/1 \tab Throw Ball Overhand \cr
   \code{dengmd024} \tab 0/1 \tab Long Jump \cr
   \code{dengmd025} \tab 0/1 \tab Balance each foot 1 second \cr
   \code{dengmd026} \tab 0/1 \tab Balance each foot 2 seconds \cr
   \code{dengmd027} \tab 0/1 \tab Hop \cr
   \code{dengmd028} \tab 0/1 \tab Balance each foot 3 seconds \cr
   \code{dengmd029} \tab 0/1 \tab Balance each foot 4 seconds \cr
   \code{dengmd030} \tab 0/1 \tab Balance each foot 5 seconds \cr
   \code{dengmd031} \tab 0/1 \tab Walk heel to toe \cr
   \code{dengmd032} \tab 0/1 \tab Balance each foot 6 seconds \cr
   \code{denlgd005} \tab 0/1 \tab Squeals \cr
   \code{denlgd006} \tab 0/1 \tab Turn to Rattle Sound \cr
   \code{denlgd007} \tab 0/1 \tab Turn to Voice \cr
   \code{denlgd008} \tab 0/1 \tab Single Syllables \cr
   \code{denlgd009} \tab 0/1 \tab Imitate Speech Sounds \cr
   \code{denlgd010} \tab 0/1 \tab Abba / Mama Nonspecific \cr
   \code{denlgd011} \tab 0/1 \tab Combine Syllables \cr
   \code{denlgd012} \tab 0/1 \tab Babbles \cr
   \code{denlgd013} \tab 0/1 \tab dada/mama specific \cr
   \code{denlgd014} \tab 0/1 \tab Say 1 Word \cr
   \code{denlgd015} \tab 0/1 \tab Say 2 Words \cr
   \code{denlgd016} \tab 0/1 \tab Say 3 Words \cr
   \code{denlgd017} \tab 0/1 \tab Say 6 Words \cr
   \code{denlgd018} \tab 0/1 \tab Point to Pictures (2) \cr
   \code{denlgd019} \tab 0/1 \tab Combine Words \cr
   \code{denlgd020} \tab 0/1 \tab Name Pictures (1) \cr
   \code{denlgd021} \tab 0/1 \tab Body Parts \cr
   \code{denlgd022} \tab 0/1 \tab Point to Pictures (4) \cr
   \code{denlgd023} \tab 0/1 \tab Speech Half Understandable \cr
   \code{denlgd024} \tab 0/1 \tab Name 4 pictures \cr
   \code{denlgd025} \tab 0/1 \tab Know Actions (2) \cr
   \code{denlgd026} \tab 0/1 \tab Know Adjectives (2) \cr
   \code{denlgd027} \tab 0/1 \tab Name Color (1 or 4) \cr
   \code{denlgd028} \tab 0/1 \tab Use of two objects \cr
   \code{denlgd029} \tab 0/1 \tab Count 1 Block \cr
   \code{denlgd030} \tab 0/1 \tab Use of four objects \cr
   \code{denlgd031} \tab 0/1 \tab Know Actions (2) (4) \cr
   \code{denlgd032} \tab 0/1 \tab Speech All-Understandable \cr
   \code{denlgd033} \tab 0/1 \tab Understand 4 Prepositions \cr
   \code{denlgd034} \tab 0/1 \tab Name Color (4) \cr
   \code{denlgd035} \tab 0/1 \tab Define 5 words \cr
   \code{denlgd036} \tab 0/1 \tab Knows adjectives (3) \cr
   \code{denlgd037} \tab 0/1 \tab Count 5 Blocks \cr
   \code{denlgd038} \tab 0/1 \tab Opposites (2) \cr
   \code{denlgd039} \tab 0/1 \tab Count blocks (5) \cr
   \code{densld005} \tab 0/1 \tab Work for toy \cr
   \code{densld006} \tab 0/1 \tab feed self \cr
   \code{densld007} \tab 0/1 \tab Clapping game "pat a cake" \cr
   \code{densld008} \tab 0/1 \tab Indicate Wants \cr
   \code{densld009} \tab 0/1 \tab wave bye-bye \cr
   \code{densld010} \tab 0/1 \tab Play Ball with You \cr
   \code{densld011} \tab 0/1 \tab Imitate Activities \cr
   \code{densld012} \tab 0/1 \tab Drink from a cup \cr
   \code{densld013} \tab 0/1 \tab help in house \cr
   \code{densld014} \tab 0/1 \tab use spoon/fork \cr
   \code{densld015} \tab 0/1 \tab remove garment \cr
   \code{densld016} \tab 0/1 \tab feed doll \cr
   \code{densld017} \tab 0/1 \tab put on clothing \cr
   \code{densld018} \tab 0/1 \tab brush teeth with help \cr
   \code{densld019} \tab 0/1 \tab wash and dry hands \cr
   \code{densld020} \tab 0/1 \tab name friend \cr
   \code{densld021} \tab 0/1 \tab put on T-shirt \cr
   \code{densld022} \tab 0/1 \tab dress no help \cr
   \code{densld023} \tab 0/1 \tab Play Games \cr
   \code{densld024} \tab 0/1 \tab brush teeth no help \cr
   \code{densld025} \tab 0/1 \tab prepare cereal \cr
   \code{mdsgmd001} \tab 0/1 \tab Sitting without support \cr
   \code{mdsgmd002} \tab 0/1 \tab Hands-and-knees crawling \cr
   \code{mdsgmd003} \tab 0/1 \tab Standing with assistance \cr
   \code{mdsgmd004} \tab 0/1 \tab Walking with assistance \cr
   \code{mdsgmd005} \tab 0/1 \tab Standing alone \cr
   \code{mdsgmd006} \tab 0/1 \tab Walking alone \cr
}
}
\usage{
gcdg_col_lt42m
}
\description{
A dataset with developmental scores at the item level for
1311 children measured at ages 0.5-3.5 year.
}
\details{
The combination of \code{subjid} and \code{agedays} identifies each unique visit (row).

Instruments: ASQ screener (\code{aqi}), Battelle (\code{bat}), Bayley III (\code{by3}),
Denver (\code{den}) and WHO motor milestones (\code{mds}).
}
\examples{
head(gcdg_col_lt42m)
}
\references{
Rubio-Codina M, Grantham-McGregor S. Predictive validity in middle
childhood of short tests of early childhood development used in large scale
studies compared to the Bayley-III, the Family Care Indicators, height-for-age,
and stunting: A longitudinal study in Bogota, Colombia.
PloS one. 2020;15:e0231317.
\url{https://journals.plos.org/plosone/article?id=10.1371/journal.pone.0231317}

Rubio-Codina M, Araujo MC, Attanasio O, Muñoz P, Grantham-McGregor S.
Concurrent validity and feasibility of short tests currently used to measure
early childhood development in large scale studies. PLoS One. 2016;11:e0160962.
\url{https://www.ncbi.nlm.nih.gov/pmc/articles/PMC4993374/}
}
\keyword{datasets}
