% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcdg_col_lt45m.R
\docType{data}
\name{gcdg_col_lt45m}
\alias{gcdg_col_lt45m}
\title{Colombia, 0.8-3.7 years: \code{gcdg_col_lt45m}}
\format{
A \code{data.frame} with 1335 rows and 269 variables:\tabular{lll}{
   Name \tab Type \tab Label \cr
   \code{ctrcd} \tab chr \tab Country code (ISO 3166-1 alpha-3) \cr
   \code{cohort} \tab chr \tab Cohort name \cr
   \code{cohortn} \tab chr \tab Cohort number \cr
   \code{id} \tab int \tab Child number \cr
   \code{agedays} \tab int \tab Age (days) \cr
   \code{sex} \tab chr \tab Either \code{"male"} or  \code{"female"} \cr
   \code{gagebrth} \tab int \tab Gestational age (days) \cr
   \code{by3cgd022} \tab 0/1 \tab Mirror Image Series: Responds positively \cr
   \code{by3cgd023} \tab 0/1 \tab Plays with string \cr
   \code{by3cgd024} \tab 0/1 \tab Bangs in play \cr
   \code{by3cgd025} \tab 0/1 \tab Searches for fallen object \cr
   \code{by3cgd026} \tab 0/1 \tab Bell Series: Manipulates \cr
   \code{by3cgd027} \tab 0/1 \tab Picks Up Block Series: Reaches for second block \cr
   \code{by3cgd028} \tab 0/1 \tab Pulls cloth to obtain object \cr
   \code{by3cgd029} \tab 0/1 \tab Pulls string adaptively \cr
   \code{by3cgd030} \tab 0/1 \tab Retains both blocks \cr
   \code{by3cgd031} \tab 0/1 \tab Bell Series: Rings purposely \cr
   \code{by3cgd032} \tab 0/1 \tab Looks at pictures \cr
   \code{by3cgd033} \tab 0/1 \tab Picks Up Block Series: Retains 2 of 3 blocks \cr
   \code{by3cgd034} \tab 0/1 \tab Searches for missing objects \cr
   \code{by3cgd035} \tab 0/1 \tab Takes blocks out of cup \cr
   \code{by3cgd036} \tab 0/1 \tab Block Series: 1 Block \cr
   \code{by3cgd037} \tab 0/1 \tab Picks up Block Series: 3 blocks \cr
   \code{by3cgd038} \tab 0/1 \tab Explores holes in pegboard \cr
   \code{by3cgd039} \tab 0/1 \tab Pushes car \cr
   \code{by3cgd040} \tab 0/1 \tab Finds hidden object \cr
   \code{by3cgd041} \tab 0/1 \tab Suspends ring \cr
   \code{by3cgd042} \tab 0/1 \tab Removes pellet \cr
   \code{by3cgd043} \tab 0/1 \tab Clear Box: Front \cr
   \code{by3cgd044} \tab 0/1 \tab Squeezes object \cr
   \code{by3cgd045} \tab 0/1 \tab Finds hidden object (Reversed) \cr
   \code{by3cgd046} \tab 0/1 \tab Removes lid from bottle \cr
   \code{by3cgd047} \tab 0/1 \tab Pegboard Series: 2 holes \cr
   \code{by3cgd048} \tab 0/1 \tab Relational Play Series: Self \cr
   \code{by3cgd049} \tab 0/1 \tab Pink Board Series: 1 piece \cr
   \code{by3cgd050} \tab 0/1 \tab Finds hidden object (Visible Displacement) \cr
   \code{by3cgd051} \tab 0/1 \tab Blue Board Series: 1 piece \cr
   \code{by3cgd052} \tab 0/1 \tab Clear Box: Sides \cr
   \code{by3cgd053} \tab 0/1 \tab Relational Play Series: Others \cr
   \code{by3cgd054} \tab 0/1 \tab Block Series: 9 Blocks \cr
   \code{by3cgd055} \tab 0/1 \tab Pegboard Series: 6 Pegs \cr
   \code{by3cgd056} \tab 0/1 \tab Pink Board Series: Completes \cr
   \code{by3cgd057} \tab 0/1 \tab Uses pencil to obtain object \cr
   \code{by3cgd058} \tab 0/1 \tab Blue Board Series: 4 Pieces \cr
   \code{by3cgd059} \tab 0/1 \tab Attends to story \cr
   \code{by3cgd060} \tab 0/1 \tab Rotated pink board \cr
   \code{by3cgd061} \tab 0/1 \tab Object assembly (Ball) \cr
   \code{by3cgd062} \tab 0/1 \tab Completes Pegboard: 25 Seconds \cr
   \code{by3cgd063} \tab 0/1 \tab Object assembly (Ice Cream Cone) \cr
   \code{by3cgd064} \tab 0/1 \tab Matches pictures \cr
   \code{by3cgd065} \tab 0/1 \tab Representational play \cr
   \code{by3cgd066} \tab 0/1 \tab Blue Board Series: Completes (75 seconds) \cr
   \code{by3cgd067} \tab 0/1 \tab Imitates a two-step action. \cr
   \code{by3cgd068} \tab 0/1 \tab Matches 3 colors \cr
   \code{by3cgd069} \tab 0/1 \tab Imaginary play \cr
   \code{by3cgd070} \tab 0/1 \tab Understands concept of one \cr
   \code{by3cgd071} \tab 0/1 \tab Multischeme combination play \cr
   \code{by3cgd072} \tab 0/1 \tab Concept Grouping: colour \cr
   \code{by3cgd073} \tab 0/1 \tab Concept Grouping: Size \cr
   \code{by3cgd074} \tab 0/1 \tab Compares masses \cr
   \code{by3cgd075} \tab 0/1 \tab Matches size \cr
   \code{by3cgd076} \tab 0/1 \tab Discriminates pictures \cr
   \code{by3cgd077} \tab 0/1 \tab Simple pattern \cr
   \code{by3cgd078} \tab 0/1 \tab Sorts pegs by colour \cr
   \code{by3cgd079} \tab 0/1 \tab Counts (One-to-one correspondence) \cr
   \code{by3cgd080} \tab 0/1 \tab Discriminates size \cr
   \code{by3cgd081} \tab 0/1 \tab Identifies 3 incomplete pictures \cr
   \code{by3cgd082} \tab 0/1 \tab Object assembly (Dog) \cr
   \code{by3cgd083} \tab 0/1 \tab Discriminates patterns \cr
   \code{by3cgd084} \tab 0/1 \tab Spatial memory \cr
   \code{by3cgd085} \tab 0/1 \tab Counts (Cardinally) \cr
   \code{by3cgd086} \tab 0/1 \tab Number constancy \cr
   \code{by3cgd087} \tab 0/1 \tab Laces card \cr
   \code{by3cgd088} \tab 0/1 \tab Classifies objects \cr
   \code{by3cgd089} \tab 0/1 \tab Understands concept of more \cr
   \code{by3exd003} \tab 0/1 \tab Vocalizes mood \cr
   \code{by3exd005} \tab 0/1 \tab Social vocalization or laughing \cr
   \code{by3exd006} \tab 0/1 \tab 2 Vowel sounds \cr
   \code{by3exd007} \tab 0/1 \tab Gets attention \cr
   \code{by3exd008} \tab 0/1 \tab 2 Consonant sounds \cr
   \code{by3exd009} \tab 0/1 \tab Uses gestures \cr
   \code{by3exd010} \tab 0/1 \tab Consonant-Vowel Combination Series: 1 Combination \cr
   \code{by3exd011} \tab 0/1 \tab Participates in play routine \cr
   \code{by3exd012} \tab 0/1 \tab Jabbers expressively \cr
   \code{by3exd013} \tab 0/1 \tab Consonant-Vowel Combination Series: 4 Combination \cr
   \code{by3exd014} \tab 0/1 \tab Uses one-word approximations \cr
   \code{by3exd015} \tab 0/1 \tab Directs attention of other \cr
   \code{by3exd016} \tab 0/1 \tab Imitates word \cr
   \code{by3exd017} \tab 0/1 \tab Initiates play interaction \cr
   \code{by3exd018} \tab 0/1 \tab Uses Words Appropriately Series: 2 words \cr
   \code{by3exd019} \tab 0/1 \tab Uses word to make wants known \cr
   \code{by3exd020} \tab 0/1 \tab Names Object Series: 1 Object \cr
   \code{by3exd021} \tab 0/1 \tab Combines word and gesture \cr
   \code{by3exd022} \tab 0/1 \tab Names Picture Series: 1 Picture \cr
   \code{by3exd023} \tab 0/1 \tab Uses Words Appropriately Series: 8 words \cr
   \code{by3exd024} \tab 0/1 \tab Answers yes or no verbally in response to questions \cr
   \code{by3exd025} \tab 0/1 \tab Imitates a two-word utterance \cr
   \code{by3exd026} \tab 0/1 \tab Uses a two-word utterance \cr
   \code{by3exd027} \tab 0/1 \tab Names Object Series: 3 Object \cr
   \code{by3exd028} \tab 0/1 \tab Names Picture Series: 5 Pictures \cr
   \code{by3exd029} \tab 0/1 \tab Uses multiple-word utterances \cr
   \code{by3exd030} \tab 0/1 \tab Uses pronouns \cr
   \code{by3exd031} \tab 0/1 \tab Names Action Picture Series: 1 Picture \cr
   \code{by3exd032} \tab 0/1 \tab Poses multiple-word questions \cr
   \code{by3exd033} \tab 0/1 \tab Makes a contingent utterance \cr
   \code{by3exd034} \tab 0/1 \tab Uses verb +ing \cr
   \code{by3exd035} \tab 0/1 \tab Names Action Picture Series: 3 Pictures \cr
   \code{by3exd036} \tab 0/1 \tab Uses different word combinations \cr
   \code{by3exd037} \tab 0/1 \tab Names Action Picture Series: 5 Pictures \cr
   \code{by3exd038} \tab 0/1 \tab Uses plurals \cr
   \code{by3exd039} \tab 0/1 \tab Answers what or where questions \cr
   \code{by3exd040} \tab 0/1 \tab Uses possessives \cr
   \code{by3exd041} \tab 0/1 \tab Names 4 colours \cr
   \code{by3exd042} \tab 0/1 \tab Answers questions logically (Related to functions) \cr
   \code{by3exd043} \tab 0/1 \tab Tells how an object is used \cr
   \code{by3exd044} \tab 0/1 \tab Uses prepositions \cr
   \code{by3exd045} \tab 0/1 \tab Uses present progressive form \cr
   \code{by3exd046} \tab 0/1 \tab Describes Pictures Series: Uses 4-5 word sentences \cr
   \code{by3exd047} \tab 0/1 \tab Describes Pictures Series: Uses past tense \cr
   \code{by3exd048} \tab 0/1 \tab Describes Pictures Series: Uses future tense \cr
   \code{by3fmd016} \tab 0/1 \tab Reaches unilaterally \cr
   \code{by3fmd017} \tab 0/1 \tab Food Pellet Series: Raking grasp \cr
   \code{by3fmd018} \tab 0/1 \tab Block Series: Partial thumb opposition \cr
   \code{by3fmd019} \tab 0/1 \tab Transfers ring \cr
   \code{by3fmd020} \tab 0/1 \tab Food Pellet Series: Whole hand grasp \cr
   \code{by3fmd021} \tab 0/1 \tab Transfers block \cr
   \code{by3fmd022} \tab 0/1 \tab Block Series: Thumb-fingertip grasp \cr
   \code{by3fmd023} \tab 0/1 \tab Brings spoons or blocks to midline \cr
   \code{by3fmd024} \tab 0/1 \tab Food Pellet Series: Partial thumb opposition \cr
   \code{by3fmd025} \tab 0/1 \tab Lifts cup by the handle \cr
   \code{by3fmd026} \tab 0/1 \tab Food Pellet Series: Thumb-fingertip grasp \cr
   \code{by3fmd027} \tab 0/1 \tab Turns pages of books \cr
   \code{by3fmd028} \tab 0/1 \tab Grasp series: Palmar grasp \cr
   \code{by3fmd029} \tab 0/1 \tab Isolates extended ring finger \cr
   \code{by3fmd030} \tab 0/1 \tab Scribbles spontaneously \cr
   \code{by3fmd031} \tab 0/1 \tab Block Stacking Series: 2 blocks \cr
   \code{by3fmd032} \tab 0/1 \tab Imitates Stroke Series: Random \cr
   \code{by3fmd033} \tab 0/1 \tab Places 10 pellets in bottle (60 seconds) \cr
   \code{by3fmd034} \tab 0/1 \tab Grasp series: Transitional grasp \cr
   \code{by3fmd035} \tab 0/1 \tab Coins in slot \cr
   \code{by3fmd036} \tab 0/1 \tab Connecting Blocks: Apart \cr
   \code{by3fmd037} \tab 0/1 \tab Grasp Series: Intermediate (Tripod) grasp \cr
   \code{by3fmd038} \tab 0/1 \tab Block stacking Series: 6 blocks \cr
   \code{by3fmd039} \tab 0/1 \tab Uses hand to hold paper in place \cr
   \code{by3fmd040} \tab 0/1 \tab Imitates Strokes Series: Horizontal \cr
   \code{by3fmd041} \tab 0/1 \tab Imitates Strokes Series: Vertical \cr
   \code{by3fmd042} \tab 0/1 \tab Connecting Blocks: Together \cr
   \code{by3fmd043} \tab 0/1 \tab Imitates Strokes Series: Circular \cr
   \code{by3fmd044} \tab 0/1 \tab Builds train of blocks \cr
   \code{by3fmd045} \tab 0/1 \tab Strings 3 blocks \cr
   \code{by3fmd046} \tab 0/1 \tab Imitates hand movements \cr
   \code{by3fmd047} \tab 0/1 \tab Snips paper \cr
   \code{by3fmd048} \tab 0/1 \tab Grasp Series: Dynamic grasp \cr
   \code{by3fmd049} \tab 0/1 \tab Tactilely discriminates shapes \cr
   \code{by3fmd050} \tab 0/1 \tab Builds wall \cr
   \code{by3fmd051} \tab 0/1 \tab Cuts paper \cr
   \code{by3fmd052} \tab 0/1 \tab Builds bridge \cr
   \code{by3fmd053} \tab 0/1 \tab Imitates plus sign \cr
   \code{by3fmd054} \tab 0/1 \tab Block stacking Series: 8 blocks \cr
   \code{by3fmd055} \tab 0/1 \tab Cuts on line \cr
   \code{by3fmd056} \tab 0/1 \tab Builds T \cr
   \code{by3fmd057} \tab 0/1 \tab Buttons 1 button \cr
   \code{by3fmd058} \tab 0/1 \tab Builds steps \cr
   \code{by3fmd059} \tab 0/1 \tab Traces designs \cr
   \code{by3fmd060} \tab 0/1 \tab Imitates square \cr
   \code{by3fmd061} \tab 0/1 \tab Copies plus sign \cr
   \code{by3fmd062} \tab 0/1 \tab Tapbs fingers \cr
   \code{by3fmd063} \tab 0/1 \tab Places 20 pellets in bottle \cr
   \code{by3fmd064} \tab 0/1 \tab Cuts circle \cr
   \code{by3fmd065} \tab 0/1 \tab Cuts square \cr
   \code{by3fmd066} \tab 0/1 \tab Copies square \cr
   \code{by3gmd019} \tab 0/1 \tab Sits with Support Series: 30 seconds \cr
   \code{by3gmd020} \tab 0/1 \tab Rolls from back to sides \cr
   \code{by3gmd021} \tab 0/1 \tab Elevates Trunk while Prone Series: Extended arms \cr
   \code{by3gmd022} \tab 0/1 \tab Sits with Support Series: 5 seconds \cr
   \code{by3gmd023} \tab 0/1 \tab Pulls up to sit \cr
   \code{by3gmd024} \tab 0/1 \tab Grasps foot with hands \cr
   \code{by3gmd025} \tab 0/1 \tab Rolls from back to stomach \cr
   \code{by3gmd026} \tab 0/1 \tab Sits without Support Series: 30 seconds \cr
   \code{by3gmd027} \tab 0/1 \tab Sits without support and holds objects \cr
   \code{by3gmd028} \tab 0/1 \tab Rotates trunk while seated \cr
   \code{by3gmd029} \tab 0/1 \tab Makes stepping movements \cr
   \code{by3gmd030} \tab 0/1 \tab Crawls Series: On stomach \cr
   \code{by3gmd031} \tab 0/1 \tab Crawls Series: Crawl position \cr
   \code{by3gmd032} \tab 0/1 \tab Moves from sitting to hands and knees \cr
   \code{by3gmd033} \tab 0/1 \tab Supports weight \cr
   \code{by3gmd034} \tab 0/1 \tab Crawls Series: Crawl movement \cr
   \code{by3gmd035} \tab 0/1 \tab Raises self to standing position \cr
   \code{by3gmd036} \tab 0/1 \tab Bounces while standing \cr
   \code{by3gmd037} \tab 0/1 \tab Walks Series: With support \cr
   \code{by3gmd038} \tab 0/1 \tab Walks sideways with support \cr
   \code{by3gmd039} \tab 0/1 \tab Sits down with control \cr
   \code{by3gmd040} \tab 0/1 \tab Stands alone \cr
   \code{by3gmd041} \tab 0/1 \tab Stands up Series: Alone \cr
   \code{by3gmd042} \tab 0/1 \tab Walks Series: Alone \cr
   \code{by3gmd043} \tab 0/1 \tab Walks Series: Alone with coordination \cr
   \code{by3gmd044} \tab 0/1 \tab Throws ball \cr
   \code{by3gmd045} \tab 0/1 \tab Squats without support \cr
   \code{by3gmd046} \tab 0/1 \tab Stands up Series: Mature \cr
   \code{by3gmd047} \tab 0/1 \tab Walks Up Stairs Series: Both feet on each step, with support \cr
   \code{by3gmd048} \tab 0/1 \tab Walks backward 2 steps \cr
   \code{by3gmd049} \tab 0/1 \tab Walks Down Stairs Series: Both feet on each step, with suppo \cr
   \code{by3gmd050} \tab 0/1 \tab Runs with coordination \cr
   \code{by3gmd051} \tab 0/1 \tab Balances on Right Foot Series: With support \cr
   \code{by3gmd052} \tab 0/1 \tab Balances on Left Foot Series: With support \cr
   \code{by3gmd053} \tab 0/1 \tab Walks sideways without support \cr
   \code{by3gmd054} \tab 0/1 \tab Jumps from bottom step \cr
   \code{by3gmd055} \tab 0/1 \tab Kicks ball \cr
   \code{by3gmd056} \tab 0/1 \tab Walks forward on path \cr
   \code{by3gmd057} \tab 0/1 \tab Walks Up Stairs Series: Both feet on each step, alone. \cr
   \code{by3gmd058} \tab 0/1 \tab Walks Down Stairs Series: Both feet on each step, alone \cr
   \code{by3gmd059} \tab 0/1 \tab Jumps Forward Series: 4 inches \cr
   \code{by3gmd060} \tab 0/1 \tab Balances on right Foot Series: 2 seconds, alone \cr
   \code{by3gmd061} \tab 0/1 \tab Balances on Left Foot Series: 2 seconds, alone \cr
   \code{by3gmd062} \tab 0/1 \tab Walks on tiptoes 4 steps \cr
   \code{by3gmd063} \tab 0/1 \tab Walks backward close to path \cr
   \code{by3gmd064} \tab 0/1 \tab Walks Up Stairs Series: Alternating feet, alone \cr
   \code{by3gmd065} \tab 0/1 \tab Imitates postures \cr
   \code{by3gmd066} \tab 0/1 \tab Stops from a full run \cr
   \code{by3gmd067} \tab 0/1 \tab Walks Down Stairs Series: Alternating feet, alone \cr
   \code{by3gmd068} \tab 0/1 \tab Hops 5 feet \cr
   \code{by3gmd069} \tab 0/1 \tab Balances on Right Foot Series: 8 seconds, alone \cr
   \code{by3gmd070} \tab 0/1 \tab Balances on Left Foot Series: 8 seconds, alone \cr
   \code{by3gmd071} \tab 0/1 \tab Walks heel to toe \cr
   \code{by3gmd072} \tab 0/1 \tab Jumps Forward Series: 24 inches \cr
   \code{by3red006} \tab 0/1 \tab Searches with head turn \cr
   \code{by3red007} \tab 0/1 \tab Discriminates sounds \cr
   \code{by3red008} \tab 0/1 \tab Sustained play with objects \cr
   \code{by3red009} \tab 0/1 \tab Responds to name \cr
   \code{by3red010} \tab 0/1 \tab Interrupts activity \cr
   \code{by3red011} \tab 0/1 \tab Recognizes 2 familiar words \cr
   \code{by3red012} \tab 0/1 \tab Responds to no-no \cr
   \code{by3red013} \tab 0/1 \tab Attends to other's play routine \cr
   \code{by3red014} \tab 0/1 \tab Responds to request for social routines \cr
   \code{by3red015} \tab 0/1 \tab Identifies Object Series: 1 correct \cr
   \code{by3red016} \tab 0/1 \tab Identifies object in the environment \cr
   \code{by3red017} \tab 0/1 \tab Identifies Picture Series: 1 correct \cr
   \code{by3red018} \tab 0/1 \tab Understands inhibitory words \cr
   \code{by3red019} \tab 0/1 \tab Identifies Object Series: 3 correct \cr
   \code{by3red020} \tab 0/1 \tab Follows one-part directions \cr
   \code{by3red021} \tab 0/1 \tab Identifies Picture Series: 3 Correct \cr
   \code{by3red022} \tab 0/1 \tab Identifies 3 clothing items \cr
   \code{by3red023} \tab 0/1 \tab Identifies Action Picture Series: 1 correct \cr
   \code{by3red024} \tab 0/1 \tab Identifies 5 parts of the body \cr
   \code{by3red025} \tab 0/1 \tab Follows two-part directions \cr
   \code{by3red026} \tab 0/1 \tab Identifies Action Picture Series: 3 correct \cr
   \code{by3red027} \tab 0/1 \tab Understands use of objects \cr
   \code{by3red028} \tab 0/1 \tab Understands part/whole relationships \cr
   \code{by3red029} \tab 0/1 \tab Identifies Action Picture Series: 5 correct \cr
   \code{by3red030} \tab 0/1 \tab Understands pronouns (him, me, my, you, your) \cr
   \code{by3red031} \tab 0/1 \tab Understands labels for sizes \cr
   \code{by3red032} \tab 0/1 \tab Understands Preposition Series: 2 correct \cr
   \code{by3red033} \tab 0/1 \tab Understands possessives \cr
   \code{by3red034} \tab 0/1 \tab Understands verb +ing \cr
   \code{by3red035} \tab 0/1 \tab Identifies colours \cr
   \code{by3red036} \tab 0/1 \tab Understands label of one \cr
   \code{by3red037} \tab 0/1 \tab Understands pronouns (they, he, she) \cr
   \code{by3red038} \tab 0/1 \tab Understands pronouns (his, her) \cr
   \code{by3red039} \tab 0/1 \tab Understands plurals \cr
   \code{by3red040} \tab 0/1 \tab Understands more \cr
   \code{by3red041} \tab 0/1 \tab Understands most \cr
   \code{by3red042} \tab 0/1 \tab Understands Preposition Series: 4 correct \cr
   \code{by3red043} \tab 0/1 \tab Understands negatives in sentences \cr
   \code{by3red044} \tab 0/1 \tab Understands past tense \cr
   \code{by3red045} \tab 0/1 \tab Understands labels for mass \cr
   \code{by3red046} \tab 0/1 \tab Understands least \cr
   \code{by3red047} \tab 0/1 \tab Understands less \cr
   \code{by3red048} \tab 0/1 \tab Understands descriptive labels \cr
   \code{by3red049} \tab 0/1 \tab Identifies categories of objects \cr
}
}
\usage{
gcdg_col_lt45m
}
\description{
A dataset with developmental scores at the item level for
707 unique children measured at two time points.
}
\details{
The combination of \code{id} and \code{agedays} identifies each unique visit (row).

Instruments: Bayley III (\code{by3}).
}
\examples{
head(gcdg_col_lt45m)
}
\references{
Attanasio OP, Fernández C, Fitzsimons EOA, Grantham-McGregor SM, Meghir C,
Rubio-Codina M. Using the infrastructure of a conditional cash transfer
program to deliver a scalable integrated early child development program in
Colombia: cluster randomized controlled trial. BMJ. 2014;349.
\url{https://www.bmj.com/content/349/bmj.g5785}
}
\keyword{datasets}
