% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcdg_ecu.R
\docType{data}
\name{gcdg_ecu}
\alias{gcdg_ecu}
\title{Ecuador, 0-3 years: \code{gcdg_ecu}}
\format{
A \code{data.frame} with 667 rows and 29 variables:\tabular{lll}{
   Name \tab Type \tab Label \cr
   \code{ctrcd} \tab chr \tab Country code (ISO 3166-1 alpha-3) \cr
   \code{cohort} \tab chr \tab Cohort name \cr
   \code{cohortn} \tab chr \tab Cohort number \cr
   \code{subjid} \tab int \tab Child number \cr
   \code{agedays} \tab int \tab Age (days) \cr
   \code{sex} \tab chr \tab Either \code{"male"} or  \code{"female"} \cr
   \code{gagebrth} \tab int \tab Gestational age (days) \cr
   \code{barxxx001} \tab 0/1 \tab Smiles in response to an adult or his/her voice \cr
   \code{barxxx002} \tab 0/1 \tab Makes sounds spontaneously or in response to a stimulus \cr
   \code{barxxx003} \tab 0/1 \tab Has head control while being pulled to stand upright \cr
   \code{barxxx004} \tab 0/1 \tab Holds a toy with one or both hands when it is presented \cr
   \code{barxxx005} \tab 0/1 \tab Rolls from back to stomach \cr
   \code{barxxx006} \tab 0/1 \tab Sits without support for some time \cr
   \code{barxxx007} \tab 0/1 \tab Starting to crawl \cr
   \code{barxxx008} \tab 0/1 \tab Grasps with fingertip to to take small objects \cr
   \code{barxxx009} \tab 0/1 \tab Stands up when being pulled \cr
   \code{barxxx010} \tab 0/1 \tab Walks with support, holding on to a railing, furniture or ad \cr
   \code{barxxx011} \tab 0/1 \tab Stands for "some moments" \cr
   \code{barxxx012} \tab 0/1 \tab Walks alone with "some steps" \cr
   \code{barxxx013} \tab 0/1 \tab Drinks well from a cup or glass \cr
   \code{barxxx014} \tab 0/1 \tab Says more than 5 words \cr
   \code{barxxx015} \tab 0/1 \tab Jumps in place \cr
   \code{barxxx016} \tab 0/1 \tab Washes and dries hands \cr
   \code{barxxx017} \tab 0/1 \tab Says his full name \cr
   \code{barxxx018} \tab 0/1 \tab Copies a circle \cr
   \code{barxxx019} \tab 0/1 \tab Balances on one foot for more than 5 seconds \cr
   \code{barxxx020} \tab 0/1 \tab Bowel and bladder control \cr
   \code{barxxx021} \tab 0/1 \tab Copies a square well \cr
   \code{barxxx022} \tab 0/1 \tab Recognizes 4 colors \cr
}
}
\usage{
gcdg_ecu
}
\description{
A dataset with developmental scores at the item level for
667 unique children. Cross-sectional.
}
\details{
The combination of \code{subjid} and \code{agedays} identifies each unique visit (row).

Instruments: Barrera Moncada (\code{bar}).
}
\references{
Paxson C, Schady N. Does money matter? The effects of cash transfers on child
development in rural Ecuador. Economic development and cultural change.
2010;59:187-229.
}
\keyword{datasets}
