% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcdg_jam_lbw.R
\docType{data}
\name{gcdg_jam_lbw}
\alias{gcdg_jam_lbw}
\title{Jamaica, low birth weight: \code{gcdg_jam_lbw}}
\format{
A \code{data.frame} with 443 rows and 192 variables:\tabular{lll}{
   Name \tab Type \tab Label \cr
   \code{ctrcd} \tab chr \tab Country code (ISO 3166-1 alpha-3) \cr
   \code{cohort} \tab chr \tab Cohort name \cr
   \code{cohortn} \tab chr \tab Cohort number \cr
   \code{subjid} \tab int \tab Child number \cr
   \code{agedays} \tab int \tab Age (days) \cr
   \code{sex} \tab chr \tab Either \code{"male"} or  \code{"female"} \cr
   \code{gagebrth} \tab int \tab Gestational age (days) \cr
   \code{gricgd012} \tab 0/1 \tab (Performance) Takes cube or toy from table \cr
   \code{gricgd013} \tab 0/1 \tab (Performance) Passes toy from hand to hand \cr
   \code{gricgd014} \tab 0/1 \tab Drops one cube for third \cr
   \code{gricgd015} \tab 0/1 \tab Manipulates 2 objects at once \cr
   \code{gricgd016} \tab 0/1 \tab Reacts to Paper IV - Plays with, crumbles, etc. \cr
   \code{gricgd017} \tab 0/1 \tab Lifts inverted cup in search of toy \cr
   \code{gricgd018} \tab 0/1 \tab Rattles box \cr
   \code{gricgd019} \tab 0/1 \tab Lifts lid off box \cr
   \code{gricgd020} \tab 0/1 \tab (Performance) Clicks 2 bricks together \cr
   \code{gricgd021} \tab 0/1 \tab Tries to take cubes out of box \cr
   \code{gricgd022} \tab 0/1 \tab Finds toy under cup \cr
   \code{gricgd023} \tab 0/1 \tab (Performance) Accepts 3rd cube without dropping \cr
   \code{gricgd024} \tab 0/1 \tab Manipulates box, lid and cubes \cr
   \code{gricgd201} \tab 0/1 \tab Remove both cubes from box (shown) \cr
   \code{gricgd202} \tab 0/1 \tab Unwraps and finds toy \cr
   \code{gricgd203} \tab 0/1 \tab One circle board - 2 trials \cr
   \code{gricgd204} \tab 0/1 \tab Opens 2 boxes \cr
   \code{gricgd205} \tab 0/1 \tab Puts cubes in and out of boxes in play \cr
   \code{gricgd206} \tab 0/1 \tab Puts 2 cubes back into one box when encouraged to do so \cr
   \code{gricgd207} \tab 0/1 \tab Two circle board - one in \cr
   \code{gricgd208} \tab 0/1 \tab Square board - 2 trials \cr
   \code{gricgd209} \tab 0/1 \tab Two circle board - two in \cr
   \code{gricgd210} \tab 0/1 \tab Can put lid back on box \cr
   \code{gricgd211} \tab 0/1 \tab (Performance) Three hole board - one in. \cr
   \code{gricgd212} \tab 0/1 \tab Puts 2 cubes into box lid on - all complete \cr
   \code{gricgd213} \tab 0/1 \tab Circle and square board together \cr
   \code{gricgd214} \tab 0/1 \tab Three-hole board - two in \cr
   \code{gricgd215} \tab 0/1 \tab (Performance) Three-hole board - three in. \cr
   \code{gricgd216} \tab 0/1 \tab Two circle board, rotated \cr
   \code{gricgd217} \tab 0/1 \tab Circle and square board rotated \cr
   \code{gricgd218} \tab 0/1 \tab (credit as 2 items) \cr
   \code{gricgd219} \tab 0/1 \tab Assembles three boxes \cr
   \code{gricgd220} \tab 0/1 \tab (credit as 2 items) \cr
   \code{gricgd221} \tab 0/1 \tab Can open screw toy \cr
   \code{gricgd222} \tab 0/1 \tab (Performance) Can open screw toy (credit as two items) \cr
   \code{gricgd223} \tab 0/1 \tab Three-hole board, rotated \cr
   \code{gricgd224} \tab 0/1 \tab (Performance) Three-hole board, rotated (credit as two items \cr
   \code{gricgd301} \tab 0/1 \tab Reassembles screw toy \cr
   \code{gricgd302} \tab 0/1 \tab Returns 9 bricks to box and replaces lid within one minute \cr
   \code{gricgd303} \tab 0/1 \tab Four-squares board completed within one minute - 2 trials \cr
   \code{gricgd304} \tab 0/1 \tab Six-hole board completed within one minute - 2 trials \cr
   \code{gricgd305} \tab 0/1 \tab Four-squares board: (40 secs). \cr
   \code{gricgd306} \tab 0/1 \tab Six-hole board: (40 secs). \cr
   \code{gricgd401} \tab 0/1 \tab Returns 9 bricks to box and replaces lid within 40 seconds \cr
   \code{gricgd402} \tab 0/1 \tab (Performance) Builds bridge with 3 boxes - inferior model bu \cr
   \code{gricgd403} \tab 0/1 \tab Assembles brick-boxes by colour - no error \cr
   \code{gricgd404} \tab 0/1 \tab Four-squares board: (15 secs) \cr
   \code{gricgd405} \tab 0/1 \tab Train under bridge successfully \cr
   \code{gricgd406} \tab 0/1 \tab Eleven-hole board (within 60 secs) Time taken (.....) \cr
   \code{griehd014} \tab 0/1 \tab Strikes one object with another \cr
   \code{griehd019} \tab 0/1 \tab Throws object \cr
   \code{griehd020} \tab 0/1 \tab (Eye and Hand Co-ordination) Thumb opposition complete. \cr
   \code{griehd021} \tab 0/1 \tab (Eye and Hand Co-ordination) Can point with index finger. \cr
   \code{griehd022} \tab 0/1 \tab Interest in motor car \cr
   \code{griehd023} \tab 0/1 \tab (Eye and Hand Co-ordination) Can hold pencil as if to mark o \cr
   \code{griehd024} \tab 0/1 \tab Likes holding little toys \cr
   \code{griehd201} \tab 0/1 \tab Uses pencil on paper a little \cr
   \code{griehd202} \tab 0/1 \tab Shows preference for one hand \cr
   \code{griehd203} \tab 0/1 \tab Plays rolling a ball \cr
   \code{griehd204} \tab 0/1 \tab Can hold 4 cubes in hands at once \cr
   \code{griehd205} \tab 0/1 \tab (Eye and Hand Co-ordination) Plays pushing little cars along \cr
   \code{griehd206} \tab 0/1 \tab Places one box, lid or brick upon another \cr
   \code{griehd207} \tab 0/1 \tab (Eye and Hand Co-ordination)  Tower of 2 bricks - imitation. \cr
   \code{griehd208} \tab 0/1 \tab (Eye and Hand Co-ordination) Pulls paper or cloth to get toy \cr
   \code{griehd209} \tab 0/1 \tab (Eye and Hand Co-ordination) Scribbles more freely, and deli \cr
   \code{griehd210} \tab 0/1 \tab Constructive play with boxes or other materials \cr
   \code{griehd211} \tab 0/1 \tab (Eye and Hand Coordination) Constructive play with boxes or \cr
   \code{griehd212} \tab 0/1 \tab Builds tower of 3 bricks \cr
   \code{griehd213} \tab 0/1 \tab Can throw a ball \cr
   \code{griehd214} \tab 0/1 \tab (Eye and hand) Can throw a ball \cr
   \code{griehd215} \tab 0/1 \tab Tower of 4+ bricks \cr
   \code{griehd216} \tab 0/1 \tab Enjoys vigourous scribble \cr
   \code{griehd217} \tab 0/1 \tab Can pour water from one cup to another \cr
   \code{griehd218} \tab 0/1 \tab Tower of 5+ bricks \cr
   \code{griehd219} \tab 0/1 \tab (Eye and Hand Co-ordination) Circular scribble in imitation \cr
   \code{griehd220} \tab 0/1 \tab Makes train of 3+ bricks \cr
   \code{griehd221} \tab 0/1 \tab (Credit as 2 items) \cr
   \code{griehd222} \tab 0/1 \tab Makes a brick or toy walk \cr
   \code{griehd223} \tab 0/1 \tab (Eye and Hand Co-ordination) Tower of 6 or 7 bricks. \cr
   \code{griehd224} \tab 0/1 \tab (Eye and Hand Co-ordination) Draws a perpendicular stroke or \cr
   \code{griehd301} \tab 0/1 \tab (Eye and Hand Co-ordination) Draws a horizontal stroke in im \cr
   \code{griehd302} \tab 0/1 \tab Threads 6 beads \cr
   \code{griehd303} \tab 0/1 \tab (Eye and Hand Co-ordination) Builds a tower of 8+ bricks. \cr
   \code{griehd304} \tab 0/1 \tab Handles scissors - tries to cut paper (4 inch square) \cr
   \code{griehd305} \tab 0/1 \tab Copies a circle - primitive model. Stage I. \cr
   \code{griehd306} \tab 0/1 \tab (Eye and Hand Co-ordination) Copies a cross -recognisable. S \cr
   \code{griehd401} \tab 0/1 \tab Folds a four-inch square of paper once (i.e. in half) \cr
   \code{griehd402} \tab 0/1 \tab Threads 12+ beads (not by pattern) \cr
   \code{griehd403} \tab 0/1 \tab (Eye and Hand Co-ordination) Can cut square into two fairly \cr
   \code{griehd404} \tab 0/1 \tab Folds a four-inch square twice \cr
   \code{griehd405} \tab 0/1 \tab Copies a 'ladder'. Stage I. \cr
   \code{griehd406} \tab 0/1 \tab Draws a 'man' recognisable. Stage I. \cr
   \code{grigmd019} \tab 0/1 \tab Crawling Reaction III: Makes some progress forwards or backw \cr
   \code{grigmd020} \tab 0/1 \tab Sits well in a chair \cr
   \code{grigmd021} \tab 0/1 \tab (Locomotor) Pulls self up and stands holding on to furniture \cr
   \code{grigmd022} \tab 0/1 \tab (Locomotor) Crawling Reaction IV: Creeps on hands and knees \cr
   \code{grigmd023} \tab 0/1 \tab (Locomotor) Side-steps around inside cot or play-pen holding \cr
   \code{grigmd024} \tab 0/1 \tab (Locomotor) Can walk when led. \cr
   \code{grigmd201} \tab 0/1 \tab Climbs on a low ledge or step \cr
   \code{grigmd202} \tab 0/1 \tab (Locomotor) Stands alone \cr
   \code{grigmd203} \tab 0/1 \tab (Locomotor) Takes a few steps alone \cr
   \code{grigmd204} \tab 0/1 \tab Kneels on floor or chair \cr
   \code{grigmd205} \tab 0/1 \tab Climbs - stairs (up) \cr
   \code{grigmd206} \tab 0/1 \tab Likes pushing pram, toy horse, etc. \cr
   \code{grigmd207} \tab 0/1 \tab (Locomotor) Walks alone well \cr
   \code{grigmd208} \tab 0/1 \tab (Locomotor) Stoops. \cr
   \code{grigmd209} \tab 0/1 \tab Develops a quick trot \cr
   \code{grigmd210} \tab 0/1 \tab Climbs into a low armchair \cr
   \code{grigmd211} \tab 0/1 \tab (Locomotor) Can walk backwards a few steps \cr
   \code{grigmd212} \tab 0/1 \tab Walks pulling toy on a string \cr
   \code{grigmd213} \tab 0/1 \tab Climbs stairs (up and down) \cr
   \code{grigmd214} \tab 0/1 \tab (Locomotor) Runs \cr
   \code{grigmd215} \tab 0/1 \tab Jumps (on the level floor) \cr
   \code{grigmd216} \tab 0/1 \tab Climbs to stand on a chair \cr
   \code{grigmd217} \tab 0/1 \tab (Locomotor) Walks upstairs- holding adult's hand \cr
   \code{grigmd218} \tab 0/1 \tab Can seat self at table \cr
   \code{grigmd219} \tab 0/1 \tab (Locomotor) Walks up and down stairs. \cr
   \code{grigmd220} \tab 0/1 \tab (Locomotor) Can kick a ball. \cr
   \code{grigmd221} \tab 0/1 \tab (Locomotor) Can jump off a step \cr
   \code{grigmd222} \tab 0/1 \tab (Locomotor) Goes alone on the stairs (any method) \cr
   \code{grigmd223} \tab 0/1 \tab Throws ball into basket \cr
   \code{grigmd224} \tab 0/1 \tab Can bring chair and seat self at table \cr
   \code{grigmd301} \tab 0/1 \tab Jumps off one step- both feet together \cr
   \code{grigmd302} \tab 0/1 \tab (Locomotor) Can stand on one foot - for 6+ seconds \cr
   \code{grigmd303} \tab 0/1 \tab Rises from kneeling without using hands \cr
   \code{grigmd304} \tab 0/1 \tab Can cross both feet and knees when seated \cr
   \code{grigmd305} \tab 0/1 \tab (Locomotor) Can stand and walk tip-toe (6+ steps) \cr
   \code{grigmd306} \tab 0/1 \tab (Locomotor) Walks upstairs: one foot on each step, adult man \cr
   \code{grigmd401} \tab 0/1 \tab Can run fast (indoors) \cr
   \code{grigmd404} \tab 0/1 \tab Walks a chalkline - 4 ft \cr
   \code{grigmd405} \tab 0/1 \tab (Locomotor) Can hop on one foot - 3+ steps \cr
   \code{grigmd406} \tab 0/1 \tab Jumps off 2 steps \cr
   \code{grigmd501} \tab 0/1 \tab Can run to kick ball \cr
   \code{grigmd502} \tab 0/1 \tab (Locomotor) Walks downstairs: one foot on each step, adult m \cr
   \code{grihsd013} \tab 0/1 \tab Responds when called \cr
   \code{grihsd014} \tab 0/1 \tab 2 syllable babble \cr
   \code{grihsd015} \tab 0/1 \tab Listens to conversations \cr
   \code{grihsd016} \tab 0/1 \tab (Hearing and Speech) Babbled phrases: 4 + syllables \cr
   \code{grihsd017} \tab 0/1 \tab (Hearing and Speech) Says Mama or Dada, etc. (one word clear \cr
   \code{grihsd018} \tab 0/1 \tab Listens to stop watch \cr
   \code{grihsd019} \tab 0/1 \tab (Hearing and Speech) Rings the bell \cr
   \code{grihsd020} \tab 0/1 \tab Shakes head for no \cr
   \code{grihsd021} \tab 0/1 \tab says 2 clear words \cr
   \code{grihsd022} \tab 0/1 \tab Short babbled sentences of 6+ syllables \cr
   \code{grihsd023} \tab 0/1 \tab Babbled monologue when alone \cr
   \code{grihsd024} \tab 0/1 \tab Says three clear words \cr
   \code{grihsd201} \tab 0/1 \tab Looks at pictures for a sew seconds \cr
   \code{grihsd202} \tab 0/1 \tab Tries definitely to sing \cr
   \code{grihsd203} \tab 0/1 \tab (Hearing and Speech)Knows own name. \cr
   \code{grihsd204} \tab 0/1 \tab Likes rhymes and jingles \cr
   \code{grihsd205} \tab 0/1 \tab (Hearing and Speech) Looks at pictures with interest \cr
   \code{grihsd206} \tab 0/1 \tab Uses 4 clear words \cr
   \code{grihsd207} \tab 0/1 \tab (Hearing and Speech) One object in box identified \cr
   \code{grihsd208} \tab 0/1 \tab Uses 5 clear words \cr
   \code{grihsd209} \tab 0/1 \tab Long babbled sentences - some words clear \cr
   \code{grihsd210} \tab 0/1 \tab Enjoys pictre book \cr
   \code{grihsd211} \tab 0/1 \tab Uses 6 or 7 clear words \cr
   \code{grihsd212} \tab 0/1 \tab Two objects in box identified \cr
   \code{grihsd213} \tab 0/1 \tab 9+  clear words \cr
   \code{grihsd214} \tab 0/1 \tab (Hearing and Speech)  Four objects in box identified \cr
   \code{grihsd215} \tab 0/1 \tab (Hearing and Speech) Picture vocabulary (1) \cr
   \code{grihsd216} \tab 0/1 \tab Uses 12+ clear words \cr
   \code{grihsd217} \tab 0/1 \tab Uses word combinations \cr
   \code{grihsd218} \tab 0/1 \tab Picture vocabulary (2) \cr
   \code{grihsd219} \tab 0/1 \tab Uses 20+ clear words \cr
   \code{grihsd220} \tab 0/1 \tab Eight objects in box identified \cr
   \code{grihsd221} \tab 0/1 \tab Listens to stories \cr
   \code{grihsd222} \tab 0/1 \tab (Hearing and Speech) Names 4 objects in box \cr
   \code{grihsd223} \tab 0/1 \tab (Hearing and Speech) Picture vocabulary (4) \cr
   \code{grihsd224} \tab 0/1 \tab (Hearing and Speech) Uses sentences of 4+ syllables \cr
   \code{grihsd301} \tab 0/1 \tab Names 12 of 18 objects in box \cr
   \code{grihsd302} \tab 0/1 \tab Picture vocabulary (12) \cr
   \code{grihsd303} \tab 0/1 \tab Defines by use (2+) - cup, knife, chair, coat, car, house, p \cr
   \code{grihsd304} \tab 0/1 \tab Repeats one six-syllable sentence \cr
   \code{grihsd305} \tab 0/1 \tab Uses 2 descriptive words \cr
   \code{grihsd306} \tab 0/1 \tab Talks well in sentences of 6+ syllables (record). \cr
   \code{grihsd401} \tab 0/1 \tab Names six or more objects in large picture \cr
   \code{grihsd402} \tab 0/1 \tab Names 17-18 objects in box \cr
   \code{grihsd403} \tab 0/1 \tab Uses 2+ personal pronouns \cr
   \code{grihsd404} \tab 0/1 \tab (Hearing and Speech) Comprehension 2+ items \cr
   \code{grihsd405} \tab 0/1 \tab Picture vocabulary (18+) \cr
   \code{grihsd406} \tab 0/1 \tab (Hearing and Speech) Knows 6+ colours \cr
   \code{grihsd501} \tab 0/1 \tab (Hearing and Speech) Defines by use: 6+. \cr
   \code{grihsd502} \tab 0/1 \tab Opposites - 2 \cr
   \code{grihsd503} \tab 0/1 \tab Materials (2+) "What is a table made of? .... a window? .... \cr
}
}
\usage{
gcdg_jam_lbw
}
\description{
A dataset with developmental scores at the item level for
226 unique children aged 1.0-2.2 years, two waves.
}
\details{
The combination of \code{subjid} and \code{agedays} identifies each unique visit (row).

Instruments: Griffiths Scales (\code{gri})
}
\references{
Walker SP, Chang SM, Powell CA, Grantham-McGregor SM. Psychosocial
intervention improves the development of term low-birth-weight infants.
The Journal of Nutrition. 2004;134:1417-1423.
}
\keyword{datasets}
