% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcdg_zaf.R
\docType{data}
\name{gcdg_zaf}
\alias{gcdg_zaf}
\title{South Africa 0.3-4.0 years: \code{gcdg_zaf}}
\format{
A \code{data.frame} with 4176 rows and 363 variables:\tabular{lll}{
   Name \tab Type \tab Label \cr
   \code{ctrcd} \tab chr \tab Country code (ISO 3166-1 alpha-3) \cr
   \code{cohort} \tab chr \tab Cohort name \cr
   \code{cohortn} \tab chr \tab Cohort number \cr
   \code{subjid} \tab int \tab Child number \cr
   \code{agedays} \tab int \tab Age (days) \cr
   \code{sex} \tab chr \tab Either \code{"male"} or  \code{"female"} \cr
   \code{gagebrth} \tab int \tab Gestational age (days) \cr
   \code{by1mdd037} \tab 0/1 \tab Reaches for dangling ring \cr
   \code{by1mdd038} \tab 0/1 \tab Pushes car along \cr
   \code{by1mdd039} \tab 0/1 \tab Fingers hand in play \cr
   \code{by1mdd040} \tab 0/1 \tab Head follows dangling ring \cr
   \code{by1mdd041} \tab 0/1 \tab Head follows vanishing spoon \cr
   \code{by1mdd042} \tab 0/1 \tab Aware of strange situation \cr
   \code{by1mdd043} \tab 0/1 \tab Manipulates table edge slightly \cr
   \code{by1mdd044} \tab 0/1 \tab Carries ring to mouth \cr
   \code{by1mdd045} \tab 0/1 \tab inspects own hand \cr
   \code{by1mdd046} \tab 0/1 \tab Closes on dangling ring (check hand preference) \cr
   \code{by1mdd047} \tab 0/1 \tab turns head to sound of bell \cr
   \code{by1mdd048} \tab 0/1 \tab turns head to sound of rattle \cr
   \code{by1mdd049} \tab 0/1 \tab reaches for cube \cr
   \code{by1mdd050} \tab 0/1 \tab manipulates table edge actively \cr
   \code{by1mdd051} \tab 0/1 \tab eye-hand coordination in reaching \cr
   \code{by1mdd052} \tab 0/1 \tab regards pellet \cr
   \code{by1mdd053} \tab 0/1 \tab mirror image approach \cr
   \code{by1mdd054} \tab 0/1 \tab picks up cube (check hand preference) \cr
   \code{by1mdd055} \tab 0/1 \tab Vocalises attitudes \cr
   \code{by1mdd056} \tab 0/1 \tab retains 2 cubes \cr
   \code{by1mdd057} \tab 0/1 \tab exploitive paper play \cr
   \code{by1mdd058} \tab 0/1 \tab Discriminates strangers \cr
   \code{by1mdd060} \tab 0/1 \tab reaches persistently \cr
   \code{by1mdd061} \tab 0/1 \tab Likes frolic play \cr
   \code{by1mdd062} \tab 0/1 \tab turns head after fallen spoon \cr
   \code{by1mdd063} \tab 0/1 \tab lifts inverted cup \cr
   \code{by1mdd064} \tab 0/1 \tab reaches for second cube \cr
   \code{by1mdd065} \tab 0/1 \tab smiles at mirror (5.4 months) \cr
   \code{by1mdd066} \tab 0/1 \tab bangs in play \cr
   \code{by1mdd067} \tab 0/1 \tab sustained inspection of ring \cr
   \code{by1mdd068} \tab 0/1 \tab exploitive string play \cr
   \code{by1mdd069} \tab 0/1 \tab transfers objects hand to hand \cr
   \code{by1mdd070} \tab 0/1 \tab picks up cube deftly and directly \cr
   \code{by1mdd071} \tab 0/1 \tab pulls string: secures ring \cr
   \code{by1mdd072} \tab 0/1 \tab interest in sound production \cr
   \code{by1mdd073} \tab 0/1 \tab lifts cup with handle \cr
   \code{by1mdd074} \tab 0/1 \tab attends to scribbling \cr
   \code{by1mdd075} \tab 0/1 \tab looks for fallen spoon \cr
   \code{by1mdd076} \tab 0/1 \tab playful response to mirror \cr
   \code{by1mdd077} \tab 0/1 \tab retains 2 of 3 cubes offered \cr
   \code{by1mdd078} \tab 0/1 \tab manipulates bell: insterest in details \cr
   \code{by1mdd079} \tab 0/1 \tab vocalizes 4 different syllables \cr
   \code{by1mdd080} \tab 0/1 \tab pulls string adaptively: secures ring \cr
   \code{by1mdd081} \tab 0/1 \tab cooperates in games (is this some other scale?) \cr
   \code{by1mdd082} \tab 0/1 \tab attempts to secure 3 cubes \cr
   \code{by1mdd083} \tab 0/1 \tab rings bell purposely \cr
   \code{by1mdd084} \tab 0/1 \tab listens selectively with familiar words \cr
   \code{by1mdd085} \tab 0/1 \tab says "da-da" or equivalent \cr
   \code{by1mdd086} \tab 0/1 \tab Uncovers toy (check manual if by pulling cloth \cr
   \code{by1mdd087} \tab 0/1 \tab fingers holes in pegboard \cr
   \code{by1mdd088} \tab 0/1 \tab picks up cup, secures cube \cr
   \code{by1mdd089} \tab 0/1 \tab responds to verbal request \cr
   \code{by1mdd090} \tab 0/1 \tab puts cube in cup on command \cr
   \code{by1mdd091} \tab 0/1 \tab looks for content of box \cr
   \code{by1mdd092} \tab 0/1 \tab stirs with spoon in imitation \cr
   \code{by1mdd093} \tab 0/1 \tab looks at pictures in book \cr
   \code{by1mdd094} \tab 0/1 \tab inhibits on command \cr
   \code{by1mdd095} \tab 0/1 \tab attempting to imitate scribble \cr
   \code{by1mdd096} \tab 0/1 \tab unwraps cube \cr
   \code{by1mdd097} \tab 0/1 \tab repeats performance laughed at \cr
   \code{by1mdd098} \tab 0/1 \tab holds crayon adaptively \cr
   \code{by1mdd099} \tab 0/1 \tab pushes car along \cr
   \code{by1mdd100} \tab 0/1 \tab puts 3 or more cubes in cup \cr
   \code{by1mdd101} \tab 0/1 \tab jabbers expressively \cr
   \code{by1mdd102} \tab 0/1 \tab uncovers blue box \cr
   \code{by1mdd103} \tab 0/1 \tab turns pages of books \cr
   \code{by1mdd104} \tab 0/1 \tab Pats toy (whistle doll) in imitation \cr
   \code{by1mdd105} \tab 0/1 \tab dangles ring by string \cr
   \code{by1mdd106} \tab 0/1 \tab imitates words \cr
   \code{by1mdd107} \tab 0/1 \tab puts beads in box (6 of 8) \cr
   \code{by1mdd108} \tab 0/1 \tab places 1 peg repeateadly \cr
   \code{by1mdd109} \tab 0/1 \tab removes pellet from bottle \cr
   \code{by1mdd110} \tab 0/1 \tab blue board: places 1 round block \cr
   \code{by1mdd111} \tab 0/1 \tab builds tower of 2 cubes \cr
   \code{by1mdd112} \tab 0/1 \tab scribbles sponstaneously \cr
   \code{by1mdd113} \tab 0/1 \tab says 2 words \cr
   \code{by1mdd114} \tab 0/1 \tab puts 9 cubes in cup \cr
   \code{by1mdd115} \tab 0/1 \tab closes round box \cr
   \code{by1mdd116} \tab 0/1 \tab uses gestures to  make wants known \cr
   \code{by1mdd117} \tab 0/1 \tab shows shoes or other clothing, or own toy \cr
   \code{by1mdd118} \tab 0/1 \tab pegs placed in 70 seconds \cr
   \code{by1mdd119} \tab 0/1 \tab builds tower of 3 cubes \cr
   \code{by1mdd120} \tab 0/1 \tab pink board: places round block \cr
   \code{by1mdd121} \tab 0/1 \tab blue board: places 2 round blocks \cr
   \code{by1mdd122} \tab 0/1 \tab attains toy with stick \cr
   \code{by1mdd123} \tab 0/1 \tab pegs placed in 42 seconds \cr
   \code{by1mdd124} \tab 0/1 \tab names 1 object \cr
   \code{by1mdd125} \tab 0/1 \tab imitates crayon stroke \cr
   \code{by1pdd016} \tab 0/1 \tab Ulnar-palmar prehension \cr
   \code{by1pdd017} \tab 0/1 \tab Sits with slight support \cr
   \code{by1pdd018} \tab 0/1 \tab Head balanced \cr
   \code{by1pdd019} \tab 0/1 \tab Turns from back to side \cr
   \code{by1pdd020} \tab 0/1 \tab Lifts head, dorsal position \cr
   \code{by1pdd021} \tab 0/1 \tab Partial thumb opposition \cr
   \code{by1pdd022} \tab 0/1 \tab Pulls to sitting \cr
   \code{by1pdd023} \tab 0/1 \tab Sits momentarily \cr
   \code{by1pdd024} \tab 0/1 \tab Unilateral reaching \cr
   \code{by1pdd025} \tab 0/1 \tab Attempts to secure pellet \cr
   \code{by1pdd026} \tab 0/1 \tab Rotates wrist \cr
   \code{by1pdd027} \tab 0/1 \tab Sits alone 30 seconds \cr
   \code{by1pdd028} \tab 0/1 \tab Rolls from back to stomach \cr
   \code{by1pdd029} \tab 0/1 \tab Sits steadily \cr
   \code{by1pdd030} \tab 0/1 \tab Scoops pellet \cr
   \code{by1pdd031} \tab 0/1 \tab Sits good coordination \cr
   \code{by1pdd032} \tab 0/1 \tab Complete thumb opposition \cr
   \code{by1pdd033} \tab 0/1 \tab Prewalking \cr
   \code{by1pdd034} \tab 0/1 \tab Early stepping movements \cr
   \code{by1pdd035} \tab 0/1 \tab Partial finger prehension \cr
   \code{by1pdd036} \tab 0/1 \tab Pulls to stand \cr
   \code{by1pdd037} \tab 0/1 \tab Raise self to sitting \cr
   \code{by1pdd038} \tab 0/1 \tab Stands up by furniture \cr
   \code{by1pdd039} \tab 0/1 \tab Combine midline \cr
   \code{by1pdd040} \tab 0/1 \tab Stepping movements \cr
   \code{by1pdd041} \tab 0/1 \tab Fine prehension \cr
   \code{by1pdd042} \tab 0/1 \tab Walks with help \cr
   \code{by1pdd043} \tab 0/1 \tab Sits down \cr
   \code{by1pdd044} \tab 0/1 \tab Play clap hands \cr
   \code{by1pdd045} \tab 0/1 \tab Stands alone \cr
   \code{by1pdd046} \tab 0/1 \tab Walks alone \cr
   \code{by1pdd047} \tab 0/1 \tab Stands up \cr
   \code{by1pdd048} \tab 0/1 \tab Throws ball \cr
   \code{by1pdd049} \tab 0/1 \tab Walsk sideways \cr
   \code{by1pdd050} \tab 0/1 \tab Walks backwards \cr
   \code{by1pdd051} \tab 0/1 \tab Stands on right foot with help \cr
   \code{by1pdd052} \tab 0/1 \tab Stands on left foot with help \cr
   \code{by1pdd053} \tab 0/1 \tab Walsk up stairs with help \cr
   \code{by1pdd054} \tab 0/1 \tab Walks down stairs with help \cr
   \code{gricgd004} \tab 0/1 \tab Holds rod put in hand (see E7) \cr
   \code{gricgd006} \tab 0/1 \tab Plays with own fingers \cr
   \code{gricgd007} \tab 0/1 \tab Resists withdrawal of rod \cr
   \code{gricgd008} \tab 0/1 \tab Claps cube put in hand \cr
   \code{gricgd009} \tab 0/1 \tab Reaches to Paper III - Pulls it away \cr
   \code{gricgd010} \tab 0/1 \tab Shows interest in box \cr
   \code{gricgd011} \tab 0/1 \tab (Performance) Holds 2 cubes. \cr
   \code{gricgd012} \tab 0/1 \tab (Performance) Takes cube or toy from table \cr
   \code{gricgd013} \tab 0/1 \tab (Performance) Passes toy from hand to hand \cr
   \code{gricgd014} \tab 0/1 \tab Drops one cube for third \cr
   \code{gricgd015} \tab 0/1 \tab Manipulates 2 objects at once \cr
   \code{gricgd016} \tab 0/1 \tab Reacts to Paper IV - Plays with, crumbles, etc. \cr
   \code{gricgd017} \tab 0/1 \tab Lifts inverted cup in search of toy \cr
   \code{gricgd018} \tab 0/1 \tab Rattles box \cr
   \code{gricgd019} \tab 0/1 \tab Lifts lid off box \cr
   \code{gricgd020} \tab 0/1 \tab (Performance) Clicks 2 bricks together \cr
   \code{gricgd021} \tab 0/1 \tab Tries to take cubes out of box \cr
   \code{gricgd022} \tab 0/1 \tab Finds toy under cup \cr
   \code{gricgd023} \tab 0/1 \tab (Performance) Accepts 3rd cube without dropping \cr
   \code{gricgd024} \tab 0/1 \tab Manipulates box, lid and cubes \cr
   \code{gricgd201} \tab 0/1 \tab Remove both cubes from box (shown) \cr
   \code{gricgd202} \tab 0/1 \tab Unwraps and finds toy \cr
   \code{gricgd203} \tab 0/1 \tab One circle board - 2 trials \cr
   \code{gricgd204} \tab 0/1 \tab Opens 2 boxes \cr
   \code{gricgd205} \tab 0/1 \tab Puts cubes in and out of boxes in play \cr
   \code{gricgd206} \tab 0/1 \tab Puts 2 cubes back into one box when encouraged to do so \cr
   \code{gricgd207} \tab 0/1 \tab Two circle board - one in \cr
   \code{gricgd208} \tab 0/1 \tab Square board - 2 trials \cr
   \code{gricgd209} \tab 0/1 \tab Two circle board - two in \cr
   \code{gricgd210} \tab 0/1 \tab Can put lid back on box \cr
   \code{gricgd211} \tab 0/1 \tab (Performance) Three hole board - one in. \cr
   \code{griehd005} \tab 0/1 \tab Follows a bell-ring - vomed in a complete circle \cr
   \code{griehd006} \tab 0/1 \tab Glances from one object to another \cr
   \code{griehd007} \tab 0/1 \tab Watches objects pulled along by string \cr
   \code{griehd008} \tab 0/1 \tab Visually explores new environment \cr
   \code{griehd009} \tab 0/1 \tab Reaches for ring and grasps \cr
   \code{griehd010} \tab 0/1 \tab Secures dangling ring \cr
   \code{griehd011} \tab 0/1 \tab Hands explore table surface \cr
   \code{griehd012} \tab 0/1 \tab Plays with ring - shaking bells, banging, etc. \cr
   \code{griehd013} \tab 0/1 \tab (Eye and Hand Co-ordination) Looks for falling object \cr
   \code{griehd014} \tab 0/1 \tab Strikes one object with another \cr
   \code{griehd015} \tab 0/1 \tab (Eye and Hand Co-ordination) Forefinger and thumb partly spe \cr
   \code{griehd016} \tab 0/1 \tab (Eye and Hand Co-ordination)  Secures ring by means of strin \cr
   \code{griehd017} \tab 0/1 \tab (Eye and Hand Co-ordination) Fine prehension. \cr
   \code{griehd018} \tab 0/1 \tab (Eye and Hand Co-ordination) Dangles ring by the string \cr
   \code{griehd019} \tab 0/1 \tab Throws object \cr
   \code{griehd020} \tab 0/1 \tab (Eye and Hand Co-ordination) Thumb opposition complete. \cr
   \code{griehd021} \tab 0/1 \tab (Eye and Hand Co-ordination) Can point with index finger. \cr
   \code{griehd022} \tab 0/1 \tab Interest in motor car \cr
   \code{griehd023} \tab 0/1 \tab (Eye and Hand Co-ordination) Can hold pencil as if to mark o \cr
   \code{griehd024} \tab 0/1 \tab Likes holding little toys \cr
   \code{griehd201} \tab 0/1 \tab Uses pencil on paper a little \cr
   \code{griehd202} \tab 0/1 \tab Shows preference for one hand \cr
   \code{griehd203} \tab 0/1 \tab Plays rolling a ball \cr
   \code{griehd204} \tab 0/1 \tab Can hold 4 cubes in hands at once \cr
   \code{griehd205} \tab 0/1 \tab (Eye and Hand Co-ordination) Plays pushing little cars along \cr
   \code{griehd206} \tab 0/1 \tab Places one box, lid or brick upon another \cr
   \code{griehd208} \tab 0/1 \tab (Eye and Hand Co-ordination) Pulls paper or cloth to get toy \cr
   \code{griehd209} \tab 0/1 \tab (Eye and Hand Co-ordination) Scribbles more freely, and deli \cr
   \code{griehd210} \tab 0/1 \tab Constructive play with boxes or other materials \cr
   \code{griehd212} \tab 0/1 \tab Builds tower of 3 bricks \cr
   \code{grigmd005} \tab 0/1 \tab Lifts head when in dorsal position \cr
   \code{grigmd006} \tab 0/1 \tab Back firm when held in siting position \cr
   \code{grigmd007} \tab 0/1 \tab Lifts head and chest when prone \cr
   \code{grigmd008} \tab 0/1 \tab Can hold head erect continuously \cr
   \code{grigmd009} \tab 0/1 \tab Lifts head and shoulders when in dorsal position \cr
   \code{grigmd010} \tab 0/1 \tab Can roll from side to side \cr
   \code{grigmd011} \tab 0/1 \tab Crawling Reaction I: Draws up knee \cr
   \code{grigmd012} \tab 0/1 \tab (Locomotor) Sits with slight support. \cr
   \code{grigmd013} \tab 0/1 \tab (Locomotor) Can roll from back to stomach, etc. \cr
   \code{grigmd014} \tab 0/1 \tab Crawling Reaction II: Tries vigorously to crawl \cr
   \code{grigmd015} \tab 0/1 \tab (Locomotor) Sits alone for a short time \cr
   \code{grigmd016} \tab 0/1 \tab (Locomotor) Stepping reaction - one foot in front of the oth \cr
   \code{grigmd017} \tab 0/1 \tab (Locomotor) Can be left sitting on the floor. \cr
   \code{grigmd018} \tab 0/1 \tab Stands when held up \cr
   \code{grigmd019} \tab 0/1 \tab Crawling Reaction III: Makes some progress forwards or backw \cr
   \code{grigmd020} \tab 0/1 \tab Sits well in a chair \cr
   \code{grigmd021} \tab 0/1 \tab (Locomotor) Pulls self up and stands holding on to furniture \cr
   \code{grigmd022} \tab 0/1 \tab (Locomotor) Crawling Reaction IV: Creeps on hands and knees \cr
   \code{grigmd023} \tab 0/1 \tab (Locomotor) Side-steps around inside cot or play-pen holding \cr
   \code{grigmd024} \tab 0/1 \tab (Locomotor) Can walk when led. \cr
   \code{grigmd201} \tab 0/1 \tab Climbs on a low ledge or step \cr
   \code{grigmd202} \tab 0/1 \tab (Locomotor) Stands alone \cr
   \code{grigmd203} \tab 0/1 \tab (Locomotor) Takes a few steps alone \cr
   \code{grigmd204} \tab 0/1 \tab Kneels on floor or chair \cr
   \code{grigmd205} \tab 0/1 \tab Climbs - stairs (up) \cr
   \code{grigmd206} \tab 0/1 \tab Likes pushing pram, toy horse, etc. \cr
   \code{grigmd207} \tab 0/1 \tab (Locomotor) Walks alone well \cr
   \code{grigmd208} \tab 0/1 \tab (Locomotor) Stoops. \cr
   \code{grigmd210} \tab 0/1 \tab Climbs into a low armchair \cr
   \code{grigmd211} \tab 0/1 \tab (Locomotor) Can walk backwards a few steps \cr
   \code{grihsd004} \tab 0/1 \tab Searches for sound with eyes \cr
   \code{grihsd005} \tab 0/1 \tab (Hearing and Speech) Makes 2 + different speech sounds \cr
   \code{grihsd006} \tab 0/1 \tab Listens to music \cr
   \code{grihsd007} \tab 0/1 \tab Searches for sound with head movements \cr
   \code{grihsd008} \tab 0/1 \tab Listens to tuning fork \cr
   \code{grihsd009} \tab 0/1 \tab (Hearing and Speech)Turns head deliberately to bell \cr
   \code{grihsd010} \tab 0/1 \tab Coos or stops crying on hearing music \cr
   \code{grihsd011} \tab 0/1 \tab Talks (babbles) to person \cr
   \code{grihsd012} \tab 0/1 \tab Makes 4+ different sounds. \cr
   \code{grihsd013} \tab 0/1 \tab Responds when called \cr
   \code{grihsd014} \tab 0/1 \tab 2 syllable babble \cr
   \code{grihsd015} \tab 0/1 \tab Listens to conversations \cr
   \code{grihsd016} \tab 0/1 \tab (Hearing and Speech) Babbled phrases: 4 + syllables \cr
   \code{grihsd017} \tab 0/1 \tab (Hearing and Speech) Says Mama or Dada, etc. (one word clear \cr
   \code{grihsd018} \tab 0/1 \tab Listens to stop watch \cr
   \code{grihsd019} \tab 0/1 \tab (Hearing and Speech) Rings the bell \cr
   \code{grihsd020} \tab 0/1 \tab Shakes head for no \cr
   \code{grihsd021} \tab 0/1 \tab says 2 clear words \cr
   \code{grihsd022} \tab 0/1 \tab Short babbled sentences of 6+ syllables \cr
   \code{grihsd023} \tab 0/1 \tab Babbled monologue when alone \cr
   \code{grihsd024} \tab 0/1 \tab Says three clear words \cr
   \code{grihsd201} \tab 0/1 \tab Looks at pictures for a sew seconds \cr
   \code{grihsd202} \tab 0/1 \tab Tries definitely to sing \cr
   \code{grihsd203} \tab 0/1 \tab (Hearing and Speech)Knows own name. \cr
   \code{grihsd204} \tab 0/1 \tab Likes rhymes and jingles \cr
   \code{grihsd206} \tab 0/1 \tab Uses 4 clear words \cr
   \code{grihsd207} \tab 0/1 \tab (Hearing and Speech) One object in box identified \cr
   \code{grihsd209} \tab 0/1 \tab Long babbled sentences - some words clear \cr
   \code{grihsd210} \tab 0/1 \tab Enjoys pictre book \cr
   \code{grihsd211} \tab 0/1 \tab Uses 6 or 7 clear words \cr
   \code{grihsd212} \tab 0/1 \tab Two objects in box identified \cr
   \code{sgrehd009} \tab 0/1 \tab Grasps ring \cr
   \code{sgrehd012} \tab 0/1 \tab Carries ring to mouth \cr
   \code{sgrehd013} \tab 0/1 \tab Clutches ring \cr
   \code{sgrehd017} \tab 0/1 \tab Reaches for string \cr
   \code{sgrehd020} \tab 0/1 \tab Watches scribble \cr
   \code{sgrehd025} \tab 0/1 \tab Pulls toy by string \cr
   \code{sgrfmd008} \tab 0/1 \tab Looks  at box on  table \cr
   \code{sgrfmd011} \tab 0/1 \tab Drops one block for second \cr
   \code{sgrfmd016} \tab 0/1 \tab Grasps box. \cr
   \code{sgrfmd017} \tab 0/1 \tab Manipulates cube \cr
   \code{sgrfmd018} \tab 0/1 \tab Takes paper \cr
   \code{sgrgmd006} \tab 0/1 \tab Active in bath \cr
   \code{sgrgmd008} \tab 0/1 \tab Rolls from side to back \cr
   \code{sgrgmd014} \tab 0/1 \tab Plays with own toes \cr
   \code{sgrgmd018} \tab 0/1 \tab Stepping movements \cr
   \code{sgrgmd022} \tab 0/1 \tab Can turn on floor \cr
   \code{sgrgmd028} \tab 0/1 \tab Can stand next to furniture \cr
   \code{sgrhsd010} \tab 0/1 \tab Laughs aloud \cr
   \code{sgrhsd015} \tab 0/1 \tab Manipulates bell \cr
   \code{sgrhsd019} \tab 0/1 \tab Shouts for attention \cr
   \code{sgrhsd021} \tab 0/1 \tab Singing tones \cr
   \code{sgrhsd029} \tab 0/1 \tab Reacts to music \cr
   \code{sgrred006} \tab 0/1 \tab Vocalises when spoken to \cr
   \code{sgrred007} \tab 0/1 \tab Follows moving person \cr
   \code{sgrred008} \tab 0/1 \tab Coos and smiles \cr
   \code{sgrred009} \tab 0/1 \tab Friendly to strangers \cr
   \code{sgrred010} \tab 0/1 \tab Resist ring being taken away \cr
   \code{sgrred011} \tab 0/1 \tab Frolics when played with \cr
   \code{sgrred012} \tab 0/1 \tab Stops crying when spoken to \cr
   \code{sgrred013} \tab 0/1 \tab Turns head to person talking \cr
   \code{sgrred014} \tab 0/1 \tab Anticipate being lifted \cr
   \code{sgrred015} \tab 0/1 \tab Holds a spoon \cr
   \code{sgrred016} \tab 0/1 \tab Stretches to be taken \cr
   \code{sgrred017} \tab 0/1 \tab Drinks from a cup \cr
   \code{sgrred018} \tab 0/1 \tab Manipulates cup or spoon in play \cr
   \code{sgrred019} \tab 0/1 \tab Reacts to mirror image \cr
   \code{sgrred020} \tab 0/1 \tab Knows strangers \cr
   \code{sgrred021} \tab 0/1 \tab Enthusiastic \cr
   \code{sgrred022} \tab 0/1 \tab Displeased when toy taken \cr
   \code{sgrred023} \tab 0/1 \tab Helps hold cup \cr
   \code{sgrred024} \tab 0/1 \tab Pulls off hat \cr
   \code{sgrred025} \tab 0/1 \tab Smiles, plays mirror \cr
   \code{sgrred026} \tab 0/1 \tab Waves bye bye \cr
   \code{sgrred027} \tab 0/1 \tab Gives affection \cr
   \code{sgrred029} \tab 0/1 \tab Plays cup, saucer \cr
   \code{sgrred030} \tab 0/1 \tab Obeys simple requests \cr
   \code{sgrred031} \tab 0/1 \tab Pays pat-a-cake \cr
   \code{sgrred032} \tab 0/1 \tab Puts block in and out of cup \cr
   \code{sgrred033} \tab 0/1 \tab Helps dressing \cr
   \code{sgrred034} \tab 0/1 \tab Hold cup \cr
   \code{sgrred035} \tab 0/1 \tab Uses spoon \cr
   \code{sgrred036} \tab 0/1 \tab Shoes shoes \cr
   \code{sgrred037} \tab 0/1 \tab tries doorknob \cr
   \code{sgrred038} \tab 0/1 \tab sks toilet \cr
   \code{sgrred039} \tab 0/1 \tab Manages cup well \cr
   \code{sgrred040} \tab 0/1 \tab Can take off shoes and socks \cr
   \code{sgrred041} \tab 0/1 \tab Like book showing \cr
   \code{vinxxc002} \tab 0/1 \tab chew solid foods \cr
   \code{vinxxc003} \tab 0/1 \tab take off socks / shoes \cr
   \code{vinxxc004} \tab 0/1 \tab transfer objects \cr
   \code{vinxxc005} \tab 0/1 \tab overcome simple obstacles \cr
   \code{vinxxc006} \tab 0/1 \tab fetch something \cr
   \code{vinxxc007} \tab 0/1 \tab drink from  cup / glass \cr
   \code{vinxxc008} \tab 0/1 \tab walk by his / herself \cr
   \code{vinxxc009} \tab 0/1 \tab get on with other children \cr
   \code{vinxxc010} \tab 0/1 \tab eat soft food with a spoon \cr
   \code{vinxxc011} \tab 0/1 \tab walk around the yard without supervision \cr
   \code{vinxxc012} \tab 0/1 \tab know what's edible \cr
   \code{vinxxc013} \tab 0/1 \tab use names of objects - shoe \cr
   \code{vinxxc014} \tab 0/1 \tab walk upstairs \cr
   \code{vinxxc015} \tab 0/1 \tab unwrap sweet \cr
   \code{vinxxc016} \tab 0/1 \tab use a short sentence \cr
   \code{vinxxc017} \tab 0/1 \tab ask or tell when it needs the toilet \cr
   \code{vinxxc018} \tab 0/1 \tab play by his / herself \cr
   \code{vinxxc019} \tab 0/1 \tab take off dress / shirt if buttons / zip undone \cr
   \code{vinxxc020} \tab 0/1 \tab get a cup of water without help \cr
   \code{vinxxc021} \tab 0/1 \tab dry hands after you have washed them \cr
   \code{vinxxc022} \tab 0/1 \tab avoid simple danger - knife / hot \cr
   \code{vinxxc023} \tab 0/1 \tab put on own dress / shirt \cr
   \code{vinxxc024} \tab 0/1 \tab tell you things / simple story \cr
   \code{vinxxc025} \tab 0/1 \tab walk downstairs \cr
   \code{vinxxc026} \tab 0/1 \tab play / do things with children - song \cr
   \code{vinxxc027} \tab 0/1 \tab put on own dress / shirt and do buttons \cr
   \code{vinxxc028} \tab 0/1 \tab help around the house / clear table \cr
   \code{vinxxc029} \tab 0/1 \tab perform for others stunts / rhymes \cr
   \code{vinxxc030} \tab 0/1 \tab wash hands and dry \cr
   \code{vinxxc031} \tab 0/1 \tab Play or do things with other children of same age eg sing so \cr
   \code{vinxxc032} \tab 0/1 \tab Use pencil or crayon for drawing \cr
   \code{vinxxc033} \tab 0/1 \tab Avoid simple dangers eg hot things, knives \cr
   \code{vinxxc034} \tab 0/1 \tab Buttons coat or dress \cr
   \code{vinxxc035} \tab 0/1 \tab Play competition exercise games eg skipping, marbles \cr
   \code{vinxxc036} \tab 0/1 \tab Walk down steps with one foot one ach step \cr
   \code{vinxxc037} \tab 0/1 \tab Dy hands without help \cr
   \code{vinxxc038} \tab 0/1 \tab Eat food with an implement \cr
   \code{vinxxc039} \tab 0/1 \tab Get a cup of water without help \cr
   \code{vinxxc040} \tab 0/1 \tab Help with little things around the house eg pick up things \cr
   \code{vinxxc041} \tab 0/1 \tab Tell things that happen or simple stories \cr
   \code{vinxxc042} \tab 0/1 \tab Put on jacket or dress without help except for zip or button \cr
   \code{vinxxc043} \tab 0/1 \tab Peform for other people, stunts, show off \cr
   \code{vinxxc044} \tab 0/1 \tab Use a pair of scissors without supervision \cr
   \code{vinxxc045} \tab 0/1 \tab Go to toilet by self, remove clothes, wipe \cr
   \code{vinxxc046} \tab 0/1 \tab Take off jacket or dress without help except for zip or butt \cr
   \code{vinxxc047} \tab 0/1 \tab Wash face without help \cr
   \code{vinxxc048} \tab 0/1 \tab Wash hands without help \cr
   \code{vinxxc049} \tab 0/1 \tab Play by themselves without supervision for short time \cr
   \code{vinxxc050} \tab 0/1 \tab Ask or tell when wants to go to toilet \cr
   \code{vinxxc051} \tab 0/1 \tab Generally dress self except for difficult fasteners \cr
}
}
\usage{
gcdg_zaf
}
\description{
A dataset with developmental scores at the item level for
2389 children measured at 4176 visits (rows) between ages 0.3-4.0 year,
with up to four visits per child.
}
\details{
The combination of \code{subjid} and \code{agedays} identifies each unique visit (row).

Instruments: Bayley I (\code{by1}), Griffiths (\code{gri}),
South African adapted Griffith (\code{sgr}), Vineland Social Maturity Scale (\code{vin})
}
\references{
Richter L, Norris S, Pettifor J, Yach D, Cameron N. Cohort profile:
Mandela’s children: the 1990 Birth to Twenty study in South Africa.
International Journal of Epidemiology. 2007;36:504-511.
\doi{10.1093/ije/dym016}
}
\keyword{datasets}
